Updated state.cpp to (almost) use flatbuffer frames. Removed frame_string references everywhere. Merged both flatbuffer IDLs.
