+++ BWEnv/include/controller.h
@@ -22,7 +22,7 @@
 #include "recorder.h"
 #include "zmq_server.h"
 
-#include "../fbs/bwenv_messages_generated.h"
+#include "../fbs/torchcraft_generated.h"
 
 class ZMQ_server;
 namespace replayer = torchcraft::replayer;
