+++ BWEnv/src/controller.cc
@@ -17,7 +17,7 @@
 #include "user_actions.h"
 #include "utils.h"
 
-#include "../fbs/bwenv_messages_generated.h"
+#include "../fbs/torchcraft_generated.h"
 
 Controller::Controller(bool is_client) {
   this->is_client = is_client;
