+++ fbs/gen.sh
@@ -1,2 +1,2 @@
 #!/bin/sh
-flatc -c --gen-mutable --scoped-enums --gen-object-api --no-includes bwenv_messages.fbs frame.fbs
+flatc -c --gen-mutable --scoped-enums --gen-object-api --no-includes torchcraft.fbs
