+++ fbs/torchcraft_generated.h
@@ -1,13 +1,45 @@
 // automatically generated by the FlatBuffers compiler, do not modify
 
 
-#ifndef FLATBUFFERS_GENERATED_FRAME_TORCHCRAFT_FBS_H_
-#define FLATBUFFERS_GENERATED_FRAME_TORCHCRAFT_FBS_H_
+#ifndef FLATBUFFERS_GENERATED_TORCHCRAFT_TORCHCRAFT_FBS_H_
+#define FLATBUFFERS_GENERATED_TORCHCRAFT_TORCHCRAFT_FBS_H_
 
 #include "flatbuffers/flatbuffers.h"
 
 namespace torchcraft {
 namespace fbs {
+
+struct Vec2;
+
+struct Command;
+struct CommandT;
+
+struct Player;
+struct PlayerT;
+
+struct HandshakeClient;
+struct HandshakeClientT;
+
+struct HandshakeServer;
+struct HandshakeServerT;
+
+struct Commands;
+struct CommandsT;
+
+struct FrameState;
+struct FrameStateT;
+
+struct PlayerLeft;
+struct PlayerLeftT;
+
+struct EndGame;
+struct EndGameT;
+
+struct Error;
+struct ErrorT;
+
+struct Message;
+struct MessageT;
 
 struct Map;
 struct MapT;
@@ -59,6 +91,333 @@ struct UnitCountT;
 struct Replayer;
 struct ReplayerT;
 
+enum class FrameRepresentation : uint8_t {
+  AsFrame = 0,
+  AsFrameDiff = 1,
+  MIN = AsFrame,
+  MAX = AsFrameDiff
+};
+
+inline FrameRepresentation (&EnumValuesFrameRepresentation())[2] {
+  static FrameRepresentation values[] = {
+    FrameRepresentation::AsFrame,
+    FrameRepresentation::AsFrameDiff
+  };
+  return values;
+}
+
+inline const char **EnumNamesFrameRepresentation() {
+  static const char *names[] = {
+    "AsFrame",
+    "AsFrameDiff",
+    nullptr
+  };
+  return names;
+}
+
+inline const char *EnumNameFrameRepresentation(FrameRepresentation e) {
+  const size_t index = static_cast<int>(e);
+  return EnumNamesFrameRepresentation()[index];
+}
+
+enum class FrameOrFrameDiff : uint8_t {
+  NONE = 0,
+  Frame = 1,
+  FrameDiff = 2,
+  MIN = NONE,
+  MAX = FrameDiff
+};
+
+inline FrameOrFrameDiff (&EnumValuesFrameOrFrameDiff())[3] {
+  static FrameOrFrameDiff values[] = {
+    FrameOrFrameDiff::NONE,
+    FrameOrFrameDiff::Frame,
+    FrameOrFrameDiff::FrameDiff
+  };
+  return values;
+}
+
+inline const char **EnumNamesFrameOrFrameDiff() {
+  static const char *names[] = {
+    "NONE",
+    "Frame",
+    "FrameDiff",
+    nullptr
+  };
+  return names;
+}
+
+inline const char *EnumNameFrameOrFrameDiff(FrameOrFrameDiff e) {
+  const size_t index = static_cast<int>(e);
+  return EnumNamesFrameOrFrameDiff()[index];
+}
+
+template<typename T> struct FrameOrFrameDiffTraits {
+  static const FrameOrFrameDiff enum_value = FrameOrFrameDiff::NONE;
+};
+
+template<> struct FrameOrFrameDiffTraits<Frame> {
+  static const FrameOrFrameDiff enum_value = FrameOrFrameDiff::Frame;
+};
+
+template<> struct FrameOrFrameDiffTraits<FrameDiff> {
+  static const FrameOrFrameDiff enum_value = FrameOrFrameDiff::FrameDiff;
+};
+
+struct FrameOrFrameDiffUnion {
+  FrameOrFrameDiff type;
+  void *value;
+
+  FrameOrFrameDiffUnion() : type(FrameOrFrameDiff::NONE), value(nullptr) {}
+  FrameOrFrameDiffUnion(FrameOrFrameDiffUnion&& u) FLATBUFFERS_NOEXCEPT :
+    type(FrameOrFrameDiff::NONE), value(nullptr)
+    { std::swap(type, u.type); std::swap(value, u.value); }
+  FrameOrFrameDiffUnion(const FrameOrFrameDiffUnion &) FLATBUFFERS_NOEXCEPT;
+  FrameOrFrameDiffUnion &operator=(const FrameOrFrameDiffUnion &u) FLATBUFFERS_NOEXCEPT
+    { FrameOrFrameDiffUnion t(u); std::swap(type, t.type); std::swap(value, t.value); return *this; }
+  FrameOrFrameDiffUnion &operator=(FrameOrFrameDiffUnion &&u) FLATBUFFERS_NOEXCEPT
+    { std::swap(type, u.type); std::swap(value, u.value); return *this; }
+  ~FrameOrFrameDiffUnion() { Reset(); }
+
+  void Reset();
+
+#ifndef FLATBUFFERS_CPP98_STL
+  template <typename T>
+  void Set(T&& val) {
+    Reset();
+    type = FrameOrFrameDiffTraits<typename T::TableType>::enum_value;
+    if (type != FrameOrFrameDiff::NONE) {
+      value = new T(std::forward<T>(val));
+    }
+  }
+#endif  // FLATBUFFERS_CPP98_STL
+
+  static void *UnPack(const void *obj, FrameOrFrameDiff type, const flatbuffers::resolver_function_t *resolver);
+  flatbuffers::Offset<void> Pack(flatbuffers::FlatBufferBuilder &_fbb, const flatbuffers::rehasher_function_t *_rehasher = nullptr) const;
+
+  FrameT *AsFrame() {
+    return type == FrameOrFrameDiff::Frame ?
+      reinterpret_cast<FrameT *>(value) : nullptr;
+  }
+  const FrameT *AsFrame() const {
+    return type == FrameOrFrameDiff::Frame ?
+      reinterpret_cast<const FrameT *>(value) : nullptr;
+  }
+  FrameDiffT *AsFrameDiff() {
+    return type == FrameOrFrameDiff::FrameDiff ?
+      reinterpret_cast<FrameDiffT *>(value) : nullptr;
+  }
+  const FrameDiffT *AsFrameDiff() const {
+    return type == FrameOrFrameDiff::FrameDiff ?
+      reinterpret_cast<const FrameDiffT *>(value) : nullptr;
+  }
+};
+
+bool VerifyFrameOrFrameDiff(flatbuffers::Verifier &verifier, const void *obj, FrameOrFrameDiff type);
+bool VerifyFrameOrFrameDiffVector(flatbuffers::Verifier &verifier, const flatbuffers::Vector<flatbuffers::Offset<void>> *values, const flatbuffers::Vector<uint8_t> *types);
+
+enum class MessageType : uint8_t {
+  NONE = 0,
+  HandshakeClient = 1,
+  Commands = 2,
+  HandshakeServer = 3,
+  FrameState = 4,
+  PlayerLeft = 5,
+  EndGame = 6,
+  Error = 7,
+  MIN = NONE,
+  MAX = Error
+};
+
+inline MessageType (&EnumValuesMessageType())[8] {
+  static MessageType values[] = {
+    MessageType::NONE,
+    MessageType::HandshakeClient,
+    MessageType::Commands,
+    MessageType::HandshakeServer,
+    MessageType::FrameState,
+    MessageType::PlayerLeft,
+    MessageType::EndGame,
+    MessageType::Error
+  };
+  return values;
+}
+
+inline const char **EnumNamesMessageType() {
+  static const char *names[] = {
+    "NONE",
+    "HandshakeClient",
+    "Commands",
+    "HandshakeServer",
+    "FrameState",
+    "PlayerLeft",
+    "EndGame",
+    "Error",
+    nullptr
+  };
+  return names;
+}
+
+inline const char *EnumNameMessageType(MessageType e) {
+  const size_t index = static_cast<int>(e);
+  return EnumNamesMessageType()[index];
+}
+
+template<typename T> struct MessageTypeTraits {
+  static const MessageType enum_value = MessageType::NONE;
+};
+
+template<> struct MessageTypeTraits<HandshakeClient> {
+  static const MessageType enum_value = MessageType::HandshakeClient;
+};
+
+template<> struct MessageTypeTraits<Commands> {
+  static const MessageType enum_value = MessageType::Commands;
+};
+
+template<> struct MessageTypeTraits<HandshakeServer> {
+  static const MessageType enum_value = MessageType::HandshakeServer;
+};
+
+template<> struct MessageTypeTraits<FrameState> {
+  static const MessageType enum_value = MessageType::FrameState;
+};
+
+template<> struct MessageTypeTraits<PlayerLeft> {
+  static const MessageType enum_value = MessageType::PlayerLeft;
+};
+
+template<> struct MessageTypeTraits<EndGame> {
+  static const MessageType enum_value = MessageType::EndGame;
+};
+
+template<> struct MessageTypeTraits<Error> {
+  static const MessageType enum_value = MessageType::Error;
+};
+
+struct MessageTypeUnion {
+  MessageType type;
+  void *value;
+
+  MessageTypeUnion() : type(MessageType::NONE), value(nullptr) {}
+  MessageTypeUnion(MessageTypeUnion&& u) FLATBUFFERS_NOEXCEPT :
+    type(MessageType::NONE), value(nullptr)
+    { std::swap(type, u.type); std::swap(value, u.value); }
+  MessageTypeUnion(const MessageTypeUnion &) FLATBUFFERS_NOEXCEPT;
+  MessageTypeUnion &operator=(const MessageTypeUnion &u) FLATBUFFERS_NOEXCEPT
+    { MessageTypeUnion t(u); std::swap(type, t.type); std::swap(value, t.value); return *this; }
+  MessageTypeUnion &operator=(MessageTypeUnion &&u) FLATBUFFERS_NOEXCEPT
+    { std::swap(type, u.type); std::swap(value, u.value); return *this; }
+  ~MessageTypeUnion() { Reset(); }
+
+  void Reset();
+
+#ifndef FLATBUFFERS_CPP98_STL
+  template <typename T>
+  void Set(T&& val) {
+    Reset();
+    type = MessageTypeTraits<typename T::TableType>::enum_value;
+    if (type != MessageType::NONE) {
+      value = new T(std::forward<T>(val));
+    }
+  }
+#endif  // FLATBUFFERS_CPP98_STL
+
+  static void *UnPack(const void *obj, MessageType type, const flatbuffers::resolver_function_t *resolver);
+  flatbuffers::Offset<void> Pack(flatbuffers::FlatBufferBuilder &_fbb, const flatbuffers::rehasher_function_t *_rehasher = nullptr) const;
+
+  HandshakeClientT *AsHandshakeClient() {
+    return type == MessageType::HandshakeClient ?
+      reinterpret_cast<HandshakeClientT *>(value) : nullptr;
+  }
+  const HandshakeClientT *AsHandshakeClient() const {
+    return type == MessageType::HandshakeClient ?
+      reinterpret_cast<const HandshakeClientT *>(value) : nullptr;
+  }
+  CommandsT *AsCommands() {
+    return type == MessageType::Commands ?
+      reinterpret_cast<CommandsT *>(value) : nullptr;
+  }
+  const CommandsT *AsCommands() const {
+    return type == MessageType::Commands ?
+      reinterpret_cast<const CommandsT *>(value) : nullptr;
+  }
+  HandshakeServerT *AsHandshakeServer() {
+    return type == MessageType::HandshakeServer ?
+      reinterpret_cast<HandshakeServerT *>(value) : nullptr;
+  }
+  const HandshakeServerT *AsHandshakeServer() const {
+    return type == MessageType::HandshakeServer ?
+      reinterpret_cast<const HandshakeServerT *>(value) : nullptr;
+  }
+  FrameStateT *AsFrameState() {
+    return type == MessageType::FrameState ?
+      reinterpret_cast<FrameStateT *>(value) : nullptr;
+  }
+  const FrameStateT *AsFrameState() const {
+    return type == MessageType::FrameState ?
+      reinterpret_cast<const FrameStateT *>(value) : nullptr;
+  }
+  PlayerLeftT *AsPlayerLeft() {
+    return type == MessageType::PlayerLeft ?
+      reinterpret_cast<PlayerLeftT *>(value) : nullptr;
+  }
+  const PlayerLeftT *AsPlayerLeft() const {
+    return type == MessageType::PlayerLeft ?
+      reinterpret_cast<const PlayerLeftT *>(value) : nullptr;
+  }
+  EndGameT *AsEndGame() {
+    return type == MessageType::EndGame ?
+      reinterpret_cast<EndGameT *>(value) : nullptr;
+  }
+  const EndGameT *AsEndGame() const {
+    return type == MessageType::EndGame ?
+      reinterpret_cast<const EndGameT *>(value) : nullptr;
+  }
+  ErrorT *AsError() {
+    return type == MessageType::Error ?
+      reinterpret_cast<ErrorT *>(value) : nullptr;
+  }
+  const ErrorT *AsError() const {
+    return type == MessageType::Error ?
+      reinterpret_cast<const ErrorT *>(value) : nullptr;
+  }
+};
+
+bool VerifyMessageType(flatbuffers::Verifier &verifier, const void *obj, MessageType type);
+bool VerifyMessageTypeVector(flatbuffers::Verifier &verifier, const flatbuffers::Vector<flatbuffers::Offset<void>> *values, const flatbuffers::Vector<uint8_t> *types);
+
+MANUALLY_ALIGNED_STRUCT(4) Vec2 FLATBUFFERS_FINAL_CLASS {
+ private:
+  int32_t x_;
+  int32_t y_;
+
+ public:
+  Vec2() {
+    memset(this, 0, sizeof(Vec2));
+  }
+  Vec2(const Vec2 &_o) {
+    memcpy(this, &_o, sizeof(Vec2));
+  }
+  Vec2(int32_t _x, int32_t _y)
+      : x_(flatbuffers::EndianScalar(_x)),
+        y_(flatbuffers::EndianScalar(_y)) {
+  }
+  int32_t x() const {
+    return flatbuffers::EndianScalar(x_);
+  }
+  void mutate_x(int32_t _x) {
+    flatbuffers::WriteScalar(&x_, _x);
+  }
+  int32_t y() const {
+    return flatbuffers::EndianScalar(y_);
+  }
+  void mutate_y(int32_t _y) {
+    flatbuffers::WriteScalar(&y_, _y);
+  }
+};
+STRUCT_END(Vec2, 8);
+
 MANUALLY_ALIGNED_STRUCT(4) Creep FLATBUFFERS_FINAL_CLASS {
  private:
   int32_t x_;
@@ -89,6 +448,1309 @@ MANUALLY_ALIGNED_STRUCT(4) Creep FLATBUFFERS_FINAL_CLASS {
   }
 };
 STRUCT_END(Creep, 8);
+
+struct CommandT : public flatbuffers::NativeTable {
+  typedef Command TableType;
+  int32_t code;
+  std::vector<int32_t> args;
+  std::string str;
+  CommandT()
+      : code(0) {
+  }
+};
+
+struct Command FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {
+  typedef CommandT NativeTableType;
+  enum {
+    VT_CODE = 4,
+    VT_ARGS = 6,
+    VT_STR = 8
+  };
+  int32_t code() const {
+    return GetField<int32_t>(VT_CODE, 0);
+  }
+  bool mutate_code(int32_t _code) {
+    return SetField<int32_t>(VT_CODE, _code, 0);
+  }
+  const flatbuffers::Vector<int32_t> *args() const {
+    return GetPointer<const flatbuffers::Vector<int32_t> *>(VT_ARGS);
+  }
+  flatbuffers::Vector<int32_t> *mutable_args() {
+    return GetPointer<flatbuffers::Vector<int32_t> *>(VT_ARGS);
+  }
+  const flatbuffers::String *str() const {
+    return GetPointer<const flatbuffers::String *>(VT_STR);
+  }
+  flatbuffers::String *mutable_str() {
+    return GetPointer<flatbuffers::String *>(VT_STR);
+  }
+  bool Verify(flatbuffers::Verifier &verifier) const {
+    return VerifyTableStart(verifier) &&
+           VerifyField<int32_t>(verifier, VT_CODE) &&
+           VerifyOffset(verifier, VT_ARGS) &&
+           verifier.Verify(args()) &&
+           VerifyOffset(verifier, VT_STR) &&
+           verifier.Verify(str()) &&
+           verifier.EndTable();
+  }
+  CommandT *UnPack(const flatbuffers::resolver_function_t *_resolver = nullptr) const;
+  void UnPackTo(CommandT *_o, const flatbuffers::resolver_function_t *_resolver = nullptr) const;
+  static flatbuffers::Offset<Command> Pack(flatbuffers::FlatBufferBuilder &_fbb, const CommandT* _o, const flatbuffers::rehasher_function_t *_rehasher = nullptr);
+};
+
+struct CommandBuilder {
+  flatbuffers::FlatBufferBuilder &fbb_;
+  flatbuffers::uoffset_t start_;
+  void add_code(int32_t code) {
+    fbb_.AddElement<int32_t>(Command::VT_CODE, code, 0);
+  }
+  void add_args(flatbuffers::Offset<flatbuffers::Vector<int32_t>> args) {
+    fbb_.AddOffset(Command::VT_ARGS, args);
+  }
+  void add_str(flatbuffers::Offset<flatbuffers::String> str) {
+    fbb_.AddOffset(Command::VT_STR, str);
+  }
+  explicit CommandBuilder(flatbuffers::FlatBufferBuilder &_fbb)
+        : fbb_(_fbb) {
+    start_ = fbb_.StartTable();
+  }
+  CommandBuilder &operator=(const CommandBuilder &);
+  flatbuffers::Offset<Command> Finish() {
+    const auto end = fbb_.EndTable(start_);
+    auto o = flatbuffers::Offset<Command>(end);
+    return o;
+  }
+};
+
+inline flatbuffers::Offset<Command> CreateCommand(
+    flatbuffers::FlatBufferBuilder &_fbb,
+    int32_t code = 0,
+    flatbuffers::Offset<flatbuffers::Vector<int32_t>> args = 0,
+    flatbuffers::Offset<flatbuffers::String> str = 0) {
+  CommandBuilder builder_(_fbb);
+  builder_.add_str(str);
+  builder_.add_args(args);
+  builder_.add_code(code);
+  return builder_.Finish();
+}
+
+inline flatbuffers::Offset<Command> CreateCommandDirect(
+    flatbuffers::FlatBufferBuilder &_fbb,
+    int32_t code = 0,
+    const std::vector<int32_t> *args = nullptr,
+    const char *str = nullptr) {
+  return torchcraft::fbs::CreateCommand(
+      _fbb,
+      code,
+      args ? _fbb.CreateVector<int32_t>(*args) : 0,
+      str ? _fbb.CreateString(str) : 0);
+}
+
+flatbuffers::Offset<Command> CreateCommand(flatbuffers::FlatBufferBuilder &_fbb, const CommandT *_o, const flatbuffers::rehasher_function_t *_rehasher = nullptr);
+
+struct PlayerT : public flatbuffers::NativeTable {
+  typedef Player TableType;
+  int32_t id;
+  int32_t race;
+  std::string name;
+  bool is_enemy;
+  PlayerT()
+      : id(0),
+        race(0),
+        is_enemy(false) {
+  }
+};
+
+struct Player FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {
+  typedef PlayerT NativeTableType;
+  enum {
+    VT_ID = 4,
+    VT_RACE = 6,
+    VT_NAME = 8,
+    VT_IS_ENEMY = 10
+  };
+  int32_t id() const {
+    return GetField<int32_t>(VT_ID, 0);
+  }
+  bool mutate_id(int32_t _id) {
+    return SetField<int32_t>(VT_ID, _id, 0);
+  }
+  int32_t race() const {
+    return GetField<int32_t>(VT_RACE, 0);
+  }
+  bool mutate_race(int32_t _race) {
+    return SetField<int32_t>(VT_RACE, _race, 0);
+  }
+  const flatbuffers::String *name() const {
+    return GetPointer<const flatbuffers::String *>(VT_NAME);
+  }
+  flatbuffers::String *mutable_name() {
+    return GetPointer<flatbuffers::String *>(VT_NAME);
+  }
+  bool is_enemy() const {
+    return GetField<uint8_t>(VT_IS_ENEMY, 0) != 0;
+  }
+  bool mutate_is_enemy(bool _is_enemy) {
+    return SetField<uint8_t>(VT_IS_ENEMY, static_cast<uint8_t>(_is_enemy), 0);
+  }
+  bool Verify(flatbuffers::Verifier &verifier) const {
+    return VerifyTableStart(verifier) &&
+           VerifyField<int32_t>(verifier, VT_ID) &&
+           VerifyField<int32_t>(verifier, VT_RACE) &&
+           VerifyOffset(verifier, VT_NAME) &&
+           verifier.Verify(name()) &&
+           VerifyField<uint8_t>(verifier, VT_IS_ENEMY) &&
+           verifier.EndTable();
+  }
+  PlayerT *UnPack(const flatbuffers::resolver_function_t *_resolver = nullptr) const;
+  void UnPackTo(PlayerT *_o, const flatbuffers::resolver_function_t *_resolver = nullptr) const;
+  static flatbuffers::Offset<Player> Pack(flatbuffers::FlatBufferBuilder &_fbb, const PlayerT* _o, const flatbuffers::rehasher_function_t *_rehasher = nullptr);
+};
+
+struct PlayerBuilder {
+  flatbuffers::FlatBufferBuilder &fbb_;
+  flatbuffers::uoffset_t start_;
+  void add_id(int32_t id) {
+    fbb_.AddElement<int32_t>(Player::VT_ID, id, 0);
+  }
+  void add_race(int32_t race) {
+    fbb_.AddElement<int32_t>(Player::VT_RACE, race, 0);
+  }
+  void add_name(flatbuffers::Offset<flatbuffers::String> name) {
+    fbb_.AddOffset(Player::VT_NAME, name);
+  }
+  void add_is_enemy(bool is_enemy) {
+    fbb_.AddElement<uint8_t>(Player::VT_IS_ENEMY, static_cast<uint8_t>(is_enemy), 0);
+  }
+  explicit PlayerBuilder(flatbuffers::FlatBufferBuilder &_fbb)
+        : fbb_(_fbb) {
+    start_ = fbb_.StartTable();
+  }
+  PlayerBuilder &operator=(const PlayerBuilder &);
+  flatbuffers::Offset<Player> Finish() {
+    const auto end = fbb_.EndTable(start_);
+    auto o = flatbuffers::Offset<Player>(end);
+    return o;
+  }
+};
+
+inline flatbuffers::Offset<Player> CreatePlayer(
+    flatbuffers::FlatBufferBuilder &_fbb,
+    int32_t id = 0,
+    int32_t race = 0,
+    flatbuffers::Offset<flatbuffers::String> name = 0,
+    bool is_enemy = false) {
+  PlayerBuilder builder_(_fbb);
+  builder_.add_name(name);
+  builder_.add_race(race);
+  builder_.add_id(id);
+  builder_.add_is_enemy(is_enemy);
+  return builder_.Finish();
+}
+
+inline flatbuffers::Offset<Player> CreatePlayerDirect(
+    flatbuffers::FlatBufferBuilder &_fbb,
+    int32_t id = 0,
+    int32_t race = 0,
+    const char *name = nullptr,
+    bool is_enemy = false) {
+  return torchcraft::fbs::CreatePlayer(
+      _fbb,
+      id,
+      race,
+      name ? _fbb.CreateString(name) : 0,
+      is_enemy);
+}
+
+flatbuffers::Offset<Player> CreatePlayer(flatbuffers::FlatBufferBuilder &_fbb, const PlayerT *_o, const flatbuffers::rehasher_function_t *_rehasher = nullptr);
+
+struct HandshakeClientT : public flatbuffers::NativeTable {
+  typedef HandshakeClient TableType;
+  int32_t protocol;
+  std::string map;
+  std::unique_ptr<Vec2> window_size;
+  std::unique_ptr<Vec2> window_pos;
+  bool micro_mode;
+  HandshakeClientT()
+      : protocol(0),
+        micro_mode(false) {
+  }
+};
+
+struct HandshakeClient FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {
+  typedef HandshakeClientT NativeTableType;
+  enum {
+    VT_PROTOCOL = 4,
+    VT_MAP = 6,
+    VT_WINDOW_SIZE = 8,
+    VT_WINDOW_POS = 10,
+    VT_MICRO_MODE = 12
+  };
+  int32_t protocol() const {
+    return GetField<int32_t>(VT_PROTOCOL, 0);
+  }
+  bool mutate_protocol(int32_t _protocol) {
+    return SetField<int32_t>(VT_PROTOCOL, _protocol, 0);
+  }
+  const flatbuffers::String *map() const {
+    return GetPointer<const flatbuffers::String *>(VT_MAP);
+  }
+  flatbuffers::String *mutable_map() {
+    return GetPointer<flatbuffers::String *>(VT_MAP);
+  }
+  const Vec2 *window_size() const {
+    return GetStruct<const Vec2 *>(VT_WINDOW_SIZE);
+  }
+  Vec2 *mutable_window_size() {
+    return GetStruct<Vec2 *>(VT_WINDOW_SIZE);
+  }
+  const Vec2 *window_pos() const {
+    return GetStruct<const Vec2 *>(VT_WINDOW_POS);
+  }
+  Vec2 *mutable_window_pos() {
+    return GetStruct<Vec2 *>(VT_WINDOW_POS);
+  }
+  bool micro_mode() const {
+    return GetField<uint8_t>(VT_MICRO_MODE, 0) != 0;
+  }
+  bool mutate_micro_mode(bool _micro_mode) {
+    return SetField<uint8_t>(VT_MICRO_MODE, static_cast<uint8_t>(_micro_mode), 0);
+  }
+  bool Verify(flatbuffers::Verifier &verifier) const {
+    return VerifyTableStart(verifier) &&
+           VerifyField<int32_t>(verifier, VT_PROTOCOL) &&
+           VerifyOffset(verifier, VT_MAP) &&
+           verifier.Verify(map()) &&
+           VerifyField<Vec2>(verifier, VT_WINDOW_SIZE) &&
+           VerifyField<Vec2>(verifier, VT_WINDOW_POS) &&
+           VerifyField<uint8_t>(verifier, VT_MICRO_MODE) &&
+           verifier.EndTable();
+  }
+  HandshakeClientT *UnPack(const flatbuffers::resolver_function_t *_resolver = nullptr) const;
+  void UnPackTo(HandshakeClientT *_o, const flatbuffers::resolver_function_t *_resolver = nullptr) const;
+  static flatbuffers::Offset<HandshakeClient> Pack(flatbuffers::FlatBufferBuilder &_fbb, const HandshakeClientT* _o, const flatbuffers::rehasher_function_t *_rehasher = nullptr);
+};
+
+struct HandshakeClientBuilder {
+  flatbuffers::FlatBufferBuilder &fbb_;
+  flatbuffers::uoffset_t start_;
+  void add_protocol(int32_t protocol) {
+    fbb_.AddElement<int32_t>(HandshakeClient::VT_PROTOCOL, protocol, 0);
+  }
+  void add_map(flatbuffers::Offset<flatbuffers::String> map) {
+    fbb_.AddOffset(HandshakeClient::VT_MAP, map);
+  }
+  void add_window_size(const Vec2 *window_size) {
+    fbb_.AddStruct(HandshakeClient::VT_WINDOW_SIZE, window_size);
+  }
+  void add_window_pos(const Vec2 *window_pos) {
+    fbb_.AddStruct(HandshakeClient::VT_WINDOW_POS, window_pos);
+  }
+  void add_micro_mode(bool micro_mode) {
+    fbb_.AddElement<uint8_t>(HandshakeClient::VT_MICRO_MODE, static_cast<uint8_t>(micro_mode), 0);
+  }
+  explicit HandshakeClientBuilder(flatbuffers::FlatBufferBuilder &_fbb)
+        : fbb_(_fbb) {
+    start_ = fbb_.StartTable();
+  }
+  HandshakeClientBuilder &operator=(const HandshakeClientBuilder &);
+  flatbuffers::Offset<HandshakeClient> Finish() {
+    const auto end = fbb_.EndTable(start_);
+    auto o = flatbuffers::Offset<HandshakeClient>(end);
+    return o;
+  }
+};
+
+inline flatbuffers::Offset<HandshakeClient> CreateHandshakeClient(
+    flatbuffers::FlatBufferBuilder &_fbb,
+    int32_t protocol = 0,
+    flatbuffers::Offset<flatbuffers::String> map = 0,
+    const Vec2 *window_size = 0,
+    const Vec2 *window_pos = 0,
+    bool micro_mode = false) {
+  HandshakeClientBuilder builder_(_fbb);
+  builder_.add_window_pos(window_pos);
+  builder_.add_window_size(window_size);
+  builder_.add_map(map);
+  builder_.add_protocol(protocol);
+  builder_.add_micro_mode(micro_mode);
+  return builder_.Finish();
+}
+
+inline flatbuffers::Offset<HandshakeClient> CreateHandshakeClientDirect(
+    flatbuffers::FlatBufferBuilder &_fbb,
+    int32_t protocol = 0,
+    const char *map = nullptr,
+    const Vec2 *window_size = 0,
+    const Vec2 *window_pos = 0,
+    bool micro_mode = false) {
+  return torchcraft::fbs::CreateHandshakeClient(
+      _fbb,
+      protocol,
+      map ? _fbb.CreateString(map) : 0,
+      window_size,
+      window_pos,
+      micro_mode);
+}
+
+flatbuffers::Offset<HandshakeClient> CreateHandshakeClient(flatbuffers::FlatBufferBuilder &_fbb, const HandshakeClientT *_o, const flatbuffers::rehasher_function_t *_rehasher = nullptr);
+
+struct HandshakeServerT : public flatbuffers::NativeTable {
+  typedef HandshakeServer TableType;
+  int32_t lag_frames;
+  std::unique_ptr<Vec2> map_size;
+  std::vector<uint8_t> ground_height_data;
+  std::vector<uint8_t> walkable_data;
+  std::string map_name;
+  bool is_replay;
+  int32_t player_id;
+  int32_t neutral_id;
+  int32_t battle_frame_count;
+  std::vector<uint8_t> buildable_data;
+  std::vector<Vec2> start_locations;
+  std::vector<std::unique_ptr<PlayerT>> players;
+  HandshakeServerT()
+      : lag_frames(0),
+        is_replay(false),
+        player_id(0),
+        neutral_id(0),
+        battle_frame_count(0) {
+  }
+};
+
+struct HandshakeServer FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {
+  typedef HandshakeServerT NativeTableType;
+  enum {
+    VT_LAG_FRAMES = 4,
+    VT_MAP_SIZE = 6,
+    VT_GROUND_HEIGHT_DATA = 8,
+    VT_WALKABLE_DATA = 10,
+    VT_MAP_NAME = 12,
+    VT_IS_REPLAY = 14,
+    VT_PLAYER_ID = 16,
+    VT_NEUTRAL_ID = 18,
+    VT_BATTLE_FRAME_COUNT = 20,
+    VT_BUILDABLE_DATA = 22,
+    VT_START_LOCATIONS = 24,
+    VT_PLAYERS = 26
+  };
+  int32_t lag_frames() const {
+    return GetField<int32_t>(VT_LAG_FRAMES, 0);
+  }
+  bool mutate_lag_frames(int32_t _lag_frames) {
+    return SetField<int32_t>(VT_LAG_FRAMES, _lag_frames, 0);
+  }
+  const Vec2 *map_size() const {
+    return GetStruct<const Vec2 *>(VT_MAP_SIZE);
+  }
+  Vec2 *mutable_map_size() {
+    return GetStruct<Vec2 *>(VT_MAP_SIZE);
+  }
+  const flatbuffers::Vector<uint8_t> *ground_height_data() const {
+    return GetPointer<const flatbuffers::Vector<uint8_t> *>(VT_GROUND_HEIGHT_DATA);
+  }
+  flatbuffers::Vector<uint8_t> *mutable_ground_height_data() {
+    return GetPointer<flatbuffers::Vector<uint8_t> *>(VT_GROUND_HEIGHT_DATA);
+  }
+  const flatbuffers::Vector<uint8_t> *walkable_data() const {
+    return GetPointer<const flatbuffers::Vector<uint8_t> *>(VT_WALKABLE_DATA);
+  }
+  flatbuffers::Vector<uint8_t> *mutable_walkable_data() {
+    return GetPointer<flatbuffers::Vector<uint8_t> *>(VT_WALKABLE_DATA);
+  }
+  const flatbuffers::String *map_name() const {
+    return GetPointer<const flatbuffers::String *>(VT_MAP_NAME);
+  }
+  flatbuffers::String *mutable_map_name() {
+    return GetPointer<flatbuffers::String *>(VT_MAP_NAME);
+  }
+  bool is_replay() const {
+    return GetField<uint8_t>(VT_IS_REPLAY, 0) != 0;
+  }
+  bool mutate_is_replay(bool _is_replay) {
+    return SetField<uint8_t>(VT_IS_REPLAY, static_cast<uint8_t>(_is_replay), 0);
+  }
+  int32_t player_id() const {
+    return GetField<int32_t>(VT_PLAYER_ID, 0);
+  }
+  bool mutate_player_id(int32_t _player_id) {
+    return SetField<int32_t>(VT_PLAYER_ID, _player_id, 0);
+  }
+  int32_t neutral_id() const {
+    return GetField<int32_t>(VT_NEUTRAL_ID, 0);
+  }
+  bool mutate_neutral_id(int32_t _neutral_id) {
+    return SetField<int32_t>(VT_NEUTRAL_ID, _neutral_id, 0);
+  }
+  int32_t battle_frame_count() const {
+    return GetField<int32_t>(VT_BATTLE_FRAME_COUNT, 0);
+  }
+  bool mutate_battle_frame_count(int32_t _battle_frame_count) {
+    return SetField<int32_t>(VT_BATTLE_FRAME_COUNT, _battle_frame_count, 0);
+  }
+  const flatbuffers::Vector<uint8_t> *buildable_data() const {
+    return GetPointer<const flatbuffers::Vector<uint8_t> *>(VT_BUILDABLE_DATA);
+  }
+  flatbuffers::Vector<uint8_t> *mutable_buildable_data() {
+    return GetPointer<flatbuffers::Vector<uint8_t> *>(VT_BUILDABLE_DATA);
+  }
+  const flatbuffers::Vector<const Vec2 *> *start_locations() const {
+    return GetPointer<const flatbuffers::Vector<const Vec2 *> *>(VT_START_LOCATIONS);
+  }
+  flatbuffers::Vector<const Vec2 *> *mutable_start_locations() {
+    return GetPointer<flatbuffers::Vector<const Vec2 *> *>(VT_START_LOCATIONS);
+  }
+  const flatbuffers::Vector<flatbuffers::Offset<Player>> *players() const {
+    return GetPointer<const flatbuffers::Vector<flatbuffers::Offset<Player>> *>(VT_PLAYERS);
+  }
+  flatbuffers::Vector<flatbuffers::Offset<Player>> *mutable_players() {
+    return GetPointer<flatbuffers::Vector<flatbuffers::Offset<Player>> *>(VT_PLAYERS);
+  }
+  bool Verify(flatbuffers::Verifier &verifier) const {
+    return VerifyTableStart(verifier) &&
+           VerifyField<int32_t>(verifier, VT_LAG_FRAMES) &&
+           VerifyField<Vec2>(verifier, VT_MAP_SIZE) &&
+           VerifyOffset(verifier, VT_GROUND_HEIGHT_DATA) &&
+           verifier.Verify(ground_height_data()) &&
+           VerifyOffset(verifier, VT_WALKABLE_DATA) &&
+           verifier.Verify(walkable_data()) &&
+           VerifyOffset(verifier, VT_MAP_NAME) &&
+           verifier.Verify(map_name()) &&
+           VerifyField<uint8_t>(verifier, VT_IS_REPLAY) &&
+           VerifyField<int32_t>(verifier, VT_PLAYER_ID) &&
+           VerifyField<int32_t>(verifier, VT_NEUTRAL_ID) &&
+           VerifyField<int32_t>(verifier, VT_BATTLE_FRAME_COUNT) &&
+           VerifyOffset(verifier, VT_BUILDABLE_DATA) &&
+           verifier.Verify(buildable_data()) &&
+           VerifyOffset(verifier, VT_START_LOCATIONS) &&
+           verifier.Verify(start_locations()) &&
+           VerifyOffset(verifier, VT_PLAYERS) &&
+           verifier.Verify(players()) &&
+           verifier.VerifyVectorOfTables(players()) &&
+           verifier.EndTable();
+  }
+  HandshakeServerT *UnPack(const flatbuffers::resolver_function_t *_resolver = nullptr) const;
+  void UnPackTo(HandshakeServerT *_o, const flatbuffers::resolver_function_t *_resolver = nullptr) const;
+  static flatbuffers::Offset<HandshakeServer> Pack(flatbuffers::FlatBufferBuilder &_fbb, const HandshakeServerT* _o, const flatbuffers::rehasher_function_t *_rehasher = nullptr);
+};
+
+struct HandshakeServerBuilder {
+  flatbuffers::FlatBufferBuilder &fbb_;
+  flatbuffers::uoffset_t start_;
+  void add_lag_frames(int32_t lag_frames) {
+    fbb_.AddElement<int32_t>(HandshakeServer::VT_LAG_FRAMES, lag_frames, 0);
+  }
+  void add_map_size(const Vec2 *map_size) {
+    fbb_.AddStruct(HandshakeServer::VT_MAP_SIZE, map_size);
+  }
+  void add_ground_height_data(flatbuffers::Offset<flatbuffers::Vector<uint8_t>> ground_height_data) {
+    fbb_.AddOffset(HandshakeServer::VT_GROUND_HEIGHT_DATA, ground_height_data);
+  }
+  void add_walkable_data(flatbuffers::Offset<flatbuffers::Vector<uint8_t>> walkable_data) {
+    fbb_.AddOffset(HandshakeServer::VT_WALKABLE_DATA, walkable_data);
+  }
+  void add_map_name(flatbuffers::Offset<flatbuffers::String> map_name) {
+    fbb_.AddOffset(HandshakeServer::VT_MAP_NAME, map_name);
+  }
+  void add_is_replay(bool is_replay) {
+    fbb_.AddElement<uint8_t>(HandshakeServer::VT_IS_REPLAY, static_cast<uint8_t>(is_replay), 0);
+  }
+  void add_player_id(int32_t player_id) {
+    fbb_.AddElement<int32_t>(HandshakeServer::VT_PLAYER_ID, player_id, 0);
+  }
+  void add_neutral_id(int32_t neutral_id) {
+    fbb_.AddElement<int32_t>(HandshakeServer::VT_NEUTRAL_ID, neutral_id, 0);
+  }
+  void add_battle_frame_count(int32_t battle_frame_count) {
+    fbb_.AddElement<int32_t>(HandshakeServer::VT_BATTLE_FRAME_COUNT, battle_frame_count, 0);
+  }
+  void add_buildable_data(flatbuffers::Offset<flatbuffers::Vector<uint8_t>> buildable_data) {
+    fbb_.AddOffset(HandshakeServer::VT_BUILDABLE_DATA, buildable_data);
+  }
+  void add_start_locations(flatbuffers::Offset<flatbuffers::Vector<const Vec2 *>> start_locations) {
+    fbb_.AddOffset(HandshakeServer::VT_START_LOCATIONS, start_locations);
+  }
+  void add_players(flatbuffers::Offset<flatbuffers::Vector<flatbuffers::Offset<Player>>> players) {
+    fbb_.AddOffset(HandshakeServer::VT_PLAYERS, players);
+  }
+  explicit HandshakeServerBuilder(flatbuffers::FlatBufferBuilder &_fbb)
+        : fbb_(_fbb) {
+    start_ = fbb_.StartTable();
+  }
+  HandshakeServerBuilder &operator=(const HandshakeServerBuilder &);
+  flatbuffers::Offset<HandshakeServer> Finish() {
+    const auto end = fbb_.EndTable(start_);
+    auto o = flatbuffers::Offset<HandshakeServer>(end);
+    return o;
+  }
+};
+
+inline flatbuffers::Offset<HandshakeServer> CreateHandshakeServer(
+    flatbuffers::FlatBufferBuilder &_fbb,
+    int32_t lag_frames = 0,
+    const Vec2 *map_size = 0,
+    flatbuffers::Offset<flatbuffers::Vector<uint8_t>> ground_height_data = 0,
+    flatbuffers::Offset<flatbuffers::Vector<uint8_t>> walkable_data = 0,
+    flatbuffers::Offset<flatbuffers::String> map_name = 0,
+    bool is_replay = false,
+    int32_t player_id = 0,
+    int32_t neutral_id = 0,
+    int32_t battle_frame_count = 0,
+    flatbuffers::Offset<flatbuffers::Vector<uint8_t>> buildable_data = 0,
+    flatbuffers::Offset<flatbuffers::Vector<const Vec2 *>> start_locations = 0,
+    flatbuffers::Offset<flatbuffers::Vector<flatbuffers::Offset<Player>>> players = 0) {
+  HandshakeServerBuilder builder_(_fbb);
+  builder_.add_players(players);
+  builder_.add_start_locations(start_locations);
+  builder_.add_buildable_data(buildable_data);
+  builder_.add_battle_frame_count(battle_frame_count);
+  builder_.add_neutral_id(neutral_id);
+  builder_.add_player_id(player_id);
+  builder_.add_map_name(map_name);
+  builder_.add_walkable_data(walkable_data);
+  builder_.add_ground_height_data(ground_height_data);
+  builder_.add_map_size(map_size);
+  builder_.add_lag_frames(lag_frames);
+  builder_.add_is_replay(is_replay);
+  return builder_.Finish();
+}
+
+inline flatbuffers::Offset<HandshakeServer> CreateHandshakeServerDirect(
+    flatbuffers::FlatBufferBuilder &_fbb,
+    int32_t lag_frames = 0,
+    const Vec2 *map_size = 0,
+    const std::vector<uint8_t> *ground_height_data = nullptr,
+    const std::vector<uint8_t> *walkable_data = nullptr,
+    const char *map_name = nullptr,
+    bool is_replay = false,
+    int32_t player_id = 0,
+    int32_t neutral_id = 0,
+    int32_t battle_frame_count = 0,
+    const std::vector<uint8_t> *buildable_data = nullptr,
+    const std::vector<const Vec2 *> *start_locations = nullptr,
+    const std::vector<flatbuffers::Offset<Player>> *players = nullptr) {
+  return torchcraft::fbs::CreateHandshakeServer(
+      _fbb,
+      lag_frames,
+      map_size,
+      ground_height_data ? _fbb.CreateVector<uint8_t>(*ground_height_data) : 0,
+      walkable_data ? _fbb.CreateVector<uint8_t>(*walkable_data) : 0,
+      map_name ? _fbb.CreateString(map_name) : 0,
+      is_replay,
+      player_id,
+      neutral_id,
+      battle_frame_count,
+      buildable_data ? _fbb.CreateVector<uint8_t>(*buildable_data) : 0,
+      start_locations ? _fbb.CreateVector<const Vec2 *>(*start_locations) : 0,
+      players ? _fbb.CreateVector<flatbuffers::Offset<Player>>(*players) : 0);
+}
+
+flatbuffers::Offset<HandshakeServer> CreateHandshakeServer(flatbuffers::FlatBufferBuilder &_fbb, const HandshakeServerT *_o, const flatbuffers::rehasher_function_t *_rehasher = nullptr);
+
+struct CommandsT : public flatbuffers::NativeTable {
+  typedef Commands TableType;
+  std::vector<std::unique_ptr<CommandT>> commands;
+  CommandsT() {
+  }
+};
+
+struct Commands FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {
+  typedef CommandsT NativeTableType;
+  enum {
+    VT_COMMANDS = 4
+  };
+  const flatbuffers::Vector<flatbuffers::Offset<Command>> *commands() const {
+    return GetPointer<const flatbuffers::Vector<flatbuffers::Offset<Command>> *>(VT_COMMANDS);
+  }
+  flatbuffers::Vector<flatbuffers::Offset<Command>> *mutable_commands() {
+    return GetPointer<flatbuffers::Vector<flatbuffers::Offset<Command>> *>(VT_COMMANDS);
+  }
+  bool Verify(flatbuffers::Verifier &verifier) const {
+    return VerifyTableStart(verifier) &&
+           VerifyOffset(verifier, VT_COMMANDS) &&
+           verifier.Verify(commands()) &&
+           verifier.VerifyVectorOfTables(commands()) &&
+           verifier.EndTable();
+  }
+  CommandsT *UnPack(const flatbuffers::resolver_function_t *_resolver = nullptr) const;
+  void UnPackTo(CommandsT *_o, const flatbuffers::resolver_function_t *_resolver = nullptr) const;
+  static flatbuffers::Offset<Commands> Pack(flatbuffers::FlatBufferBuilder &_fbb, const CommandsT* _o, const flatbuffers::rehasher_function_t *_rehasher = nullptr);
+};
+
+struct CommandsBuilder {
+  flatbuffers::FlatBufferBuilder &fbb_;
+  flatbuffers::uoffset_t start_;
+  void add_commands(flatbuffers::Offset<flatbuffers::Vector<flatbuffers::Offset<Command>>> commands) {
+    fbb_.AddOffset(Commands::VT_COMMANDS, commands);
+  }
+  explicit CommandsBuilder(flatbuffers::FlatBufferBuilder &_fbb)
+        : fbb_(_fbb) {
+    start_ = fbb_.StartTable();
+  }
+  CommandsBuilder &operator=(const CommandsBuilder &);
+  flatbuffers::Offset<Commands> Finish() {
+    const auto end = fbb_.EndTable(start_);
+    auto o = flatbuffers::Offset<Commands>(end);
+    return o;
+  }
+};
+
+inline flatbuffers::Offset<Commands> CreateCommands(
+    flatbuffers::FlatBufferBuilder &_fbb,
+    flatbuffers::Offset<flatbuffers::Vector<flatbuffers::Offset<Command>>> commands = 0) {
+  CommandsBuilder builder_(_fbb);
+  builder_.add_commands(commands);
+  return builder_.Finish();
+}
+
+inline flatbuffers::Offset<Commands> CreateCommandsDirect(
+    flatbuffers::FlatBufferBuilder &_fbb,
+    const std::vector<flatbuffers::Offset<Command>> *commands = nullptr) {
+  return torchcraft::fbs::CreateCommands(
+      _fbb,
+      commands ? _fbb.CreateVector<flatbuffers::Offset<Command>>(*commands) : 0);
+}
+
+flatbuffers::Offset<Commands> CreateCommands(flatbuffers::FlatBufferBuilder &_fbb, const CommandsT *_o, const flatbuffers::rehasher_function_t *_rehasher = nullptr);
+
+struct FrameStateT : public flatbuffers::NativeTable {
+  typedef FrameState TableType;
+  FrameOrFrameDiffUnion frameOrFrameDiff;
+  std::vector<int32_t> deaths;
+  int32_t frame_from_bwapi;
+  int32_t battle_frame_count;
+  std::vector<int8_t> commands_status;
+  std::string img_mode;
+  std::unique_ptr<Vec2> screen_position;
+  std::vector<uint8_t> visibility;
+  std::unique_ptr<Vec2> visibility_size;
+  std::vector<uint8_t> img_data;
+  std::unique_ptr<Vec2> img_size;
+  FrameStateT()
+      : frame_from_bwapi(0),
+        battle_frame_count(0) {
+  }
+};
+
+struct FrameState FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {
+  typedef FrameStateT NativeTableType;
+  enum {
+    VT_FRAMEORFRAMEDIFF_TYPE = 4,
+    VT_FRAMEORFRAMEDIFF = 6,
+    VT_DEATHS = 8,
+    VT_FRAME_FROM_BWAPI = 10,
+    VT_BATTLE_FRAME_COUNT = 12,
+    VT_COMMANDS_STATUS = 14,
+    VT_IMG_MODE = 16,
+    VT_SCREEN_POSITION = 18,
+    VT_VISIBILITY = 20,
+    VT_VISIBILITY_SIZE = 22,
+    VT_IMG_DATA = 24,
+    VT_IMG_SIZE = 26
+  };
+  FrameOrFrameDiff frameOrFrameDiff_type() const {
+    return static_cast<FrameOrFrameDiff>(GetField<uint8_t>(VT_FRAMEORFRAMEDIFF_TYPE, 0));
+  }
+  bool mutate_frameOrFrameDiff_type(FrameOrFrameDiff _frameOrFrameDiff_type) {
+    return SetField<uint8_t>(VT_FRAMEORFRAMEDIFF_TYPE, static_cast<uint8_t>(_frameOrFrameDiff_type), 0);
+  }
+  const void *frameOrFrameDiff() const {
+    return GetPointer<const void *>(VT_FRAMEORFRAMEDIFF);
+  }
+  template<typename T> const T *frameOrFrameDiff_as() const;
+  const Frame *frameOrFrameDiff_as_Frame() const {
+    return frameOrFrameDiff_type() == FrameOrFrameDiff::Frame ? static_cast<const Frame *>(frameOrFrameDiff()) : nullptr;
+  }
+  const FrameDiff *frameOrFrameDiff_as_FrameDiff() const {
+    return frameOrFrameDiff_type() == FrameOrFrameDiff::FrameDiff ? static_cast<const FrameDiff *>(frameOrFrameDiff()) : nullptr;
+  }
+  void *mutable_frameOrFrameDiff() {
+    return GetPointer<void *>(VT_FRAMEORFRAMEDIFF);
+  }
+  const flatbuffers::Vector<int32_t> *deaths() const {
+    return GetPointer<const flatbuffers::Vector<int32_t> *>(VT_DEATHS);
+  }
+  flatbuffers::Vector<int32_t> *mutable_deaths() {
+    return GetPointer<flatbuffers::Vector<int32_t> *>(VT_DEATHS);
+  }
+  int32_t frame_from_bwapi() const {
+    return GetField<int32_t>(VT_FRAME_FROM_BWAPI, 0);
+  }
+  bool mutate_frame_from_bwapi(int32_t _frame_from_bwapi) {
+    return SetField<int32_t>(VT_FRAME_FROM_BWAPI, _frame_from_bwapi, 0);
+  }
+  int32_t battle_frame_count() const {
+    return GetField<int32_t>(VT_BATTLE_FRAME_COUNT, 0);
+  }
+  bool mutate_battle_frame_count(int32_t _battle_frame_count) {
+    return SetField<int32_t>(VT_BATTLE_FRAME_COUNT, _battle_frame_count, 0);
+  }
+  const flatbuffers::Vector<int8_t> *commands_status() const {
+    return GetPointer<const flatbuffers::Vector<int8_t> *>(VT_COMMANDS_STATUS);
+  }
+  flatbuffers::Vector<int8_t> *mutable_commands_status() {
+    return GetPointer<flatbuffers::Vector<int8_t> *>(VT_COMMANDS_STATUS);
+  }
+  const flatbuffers::String *img_mode() const {
+    return GetPointer<const flatbuffers::String *>(VT_IMG_MODE);
+  }
+  flatbuffers::String *mutable_img_mode() {
+    return GetPointer<flatbuffers::String *>(VT_IMG_MODE);
+  }
+  const Vec2 *screen_position() const {
+    return GetStruct<const Vec2 *>(VT_SCREEN_POSITION);
+  }
+  Vec2 *mutable_screen_position() {
+    return GetStruct<Vec2 *>(VT_SCREEN_POSITION);
+  }
+  const flatbuffers::Vector<uint8_t> *visibility() const {
+    return GetPointer<const flatbuffers::Vector<uint8_t> *>(VT_VISIBILITY);
+  }
+  flatbuffers::Vector<uint8_t> *mutable_visibility() {
+    return GetPointer<flatbuffers::Vector<uint8_t> *>(VT_VISIBILITY);
+  }
+  const Vec2 *visibility_size() const {
+    return GetStruct<const Vec2 *>(VT_VISIBILITY_SIZE);
+  }
+  Vec2 *mutable_visibility_size() {
+    return GetStruct<Vec2 *>(VT_VISIBILITY_SIZE);
+  }
+  const flatbuffers::Vector<uint8_t> *img_data() const {
+    return GetPointer<const flatbuffers::Vector<uint8_t> *>(VT_IMG_DATA);
+  }
+  flatbuffers::Vector<uint8_t> *mutable_img_data() {
+    return GetPointer<flatbuffers::Vector<uint8_t> *>(VT_IMG_DATA);
+  }
+  const Vec2 *img_size() const {
+    return GetStruct<const Vec2 *>(VT_IMG_SIZE);
+  }
+  Vec2 *mutable_img_size() {
+    return GetStruct<Vec2 *>(VT_IMG_SIZE);
+  }
+  bool Verify(flatbuffers::Verifier &verifier) const {
+    return VerifyTableStart(verifier) &&
+           VerifyField<uint8_t>(verifier, VT_FRAMEORFRAMEDIFF_TYPE) &&
+           VerifyOffset(verifier, VT_FRAMEORFRAMEDIFF) &&
+           VerifyFrameOrFrameDiff(verifier, frameOrFrameDiff(), frameOrFrameDiff_type()) &&
+           VerifyOffset(verifier, VT_DEATHS) &&
+           verifier.Verify(deaths()) &&
+           VerifyField<int32_t>(verifier, VT_FRAME_FROM_BWAPI) &&
+           VerifyField<int32_t>(verifier, VT_BATTLE_FRAME_COUNT) &&
+           VerifyOffset(verifier, VT_COMMANDS_STATUS) &&
+           verifier.Verify(commands_status()) &&
+           VerifyOffset(verifier, VT_IMG_MODE) &&
+           verifier.Verify(img_mode()) &&
+           VerifyField<Vec2>(verifier, VT_SCREEN_POSITION) &&
+           VerifyOffset(verifier, VT_VISIBILITY) &&
+           verifier.Verify(visibility()) &&
+           VerifyField<Vec2>(verifier, VT_VISIBILITY_SIZE) &&
+           VerifyOffset(verifier, VT_IMG_DATA) &&
+           verifier.Verify(img_data()) &&
+           VerifyField<Vec2>(verifier, VT_IMG_SIZE) &&
+           verifier.EndTable();
+  }
+  FrameStateT *UnPack(const flatbuffers::resolver_function_t *_resolver = nullptr) const;
+  void UnPackTo(FrameStateT *_o, const flatbuffers::resolver_function_t *_resolver = nullptr) const;
+  static flatbuffers::Offset<FrameState> Pack(flatbuffers::FlatBufferBuilder &_fbb, const FrameStateT* _o, const flatbuffers::rehasher_function_t *_rehasher = nullptr);
+};
+
+template<> inline const Frame *FrameState::frameOrFrameDiff_as<Frame>() const {
+  return frameOrFrameDiff_as_Frame();
+}
+
+template<> inline const FrameDiff *FrameState::frameOrFrameDiff_as<FrameDiff>() const {
+  return frameOrFrameDiff_as_FrameDiff();
+}
+
+struct FrameStateBuilder {
+  flatbuffers::FlatBufferBuilder &fbb_;
+  flatbuffers::uoffset_t start_;
+  void add_frameOrFrameDiff_type(FrameOrFrameDiff frameOrFrameDiff_type) {
+    fbb_.AddElement<uint8_t>(FrameState::VT_FRAMEORFRAMEDIFF_TYPE, static_cast<uint8_t>(frameOrFrameDiff_type), 0);
+  }
+  void add_frameOrFrameDiff(flatbuffers::Offset<void> frameOrFrameDiff) {
+    fbb_.AddOffset(FrameState::VT_FRAMEORFRAMEDIFF, frameOrFrameDiff);
+  }
+  void add_deaths(flatbuffers::Offset<flatbuffers::Vector<int32_t>> deaths) {
+    fbb_.AddOffset(FrameState::VT_DEATHS, deaths);
+  }
+  void add_frame_from_bwapi(int32_t frame_from_bwapi) {
+    fbb_.AddElement<int32_t>(FrameState::VT_FRAME_FROM_BWAPI, frame_from_bwapi, 0);
+  }
+  void add_battle_frame_count(int32_t battle_frame_count) {
+    fbb_.AddElement<int32_t>(FrameState::VT_BATTLE_FRAME_COUNT, battle_frame_count, 0);
+  }
+  void add_commands_status(flatbuffers::Offset<flatbuffers::Vector<int8_t>> commands_status) {
+    fbb_.AddOffset(FrameState::VT_COMMANDS_STATUS, commands_status);
+  }
+  void add_img_mode(flatbuffers::Offset<flatbuffers::String> img_mode) {
+    fbb_.AddOffset(FrameState::VT_IMG_MODE, img_mode);
+  }
+  void add_screen_position(const Vec2 *screen_position) {
+    fbb_.AddStruct(FrameState::VT_SCREEN_POSITION, screen_position);
+  }
+  void add_visibility(flatbuffers::Offset<flatbuffers::Vector<uint8_t>> visibility) {
+    fbb_.AddOffset(FrameState::VT_VISIBILITY, visibility);
+  }
+  void add_visibility_size(const Vec2 *visibility_size) {
+    fbb_.AddStruct(FrameState::VT_VISIBILITY_SIZE, visibility_size);
+  }
+  void add_img_data(flatbuffers::Offset<flatbuffers::Vector<uint8_t>> img_data) {
+    fbb_.AddOffset(FrameState::VT_IMG_DATA, img_data);
+  }
+  void add_img_size(const Vec2 *img_size) {
+    fbb_.AddStruct(FrameState::VT_IMG_SIZE, img_size);
+  }
+  explicit FrameStateBuilder(flatbuffers::FlatBufferBuilder &_fbb)
+        : fbb_(_fbb) {
+    start_ = fbb_.StartTable();
+  }
+  FrameStateBuilder &operator=(const FrameStateBuilder &);
+  flatbuffers::Offset<FrameState> Finish() {
+    const auto end = fbb_.EndTable(start_);
+    auto o = flatbuffers::Offset<FrameState>(end);
+    return o;
+  }
+};
+
+inline flatbuffers::Offset<FrameState> CreateFrameState(
+    flatbuffers::FlatBufferBuilder &_fbb,
+    FrameOrFrameDiff frameOrFrameDiff_type = FrameOrFrameDiff::NONE,
+    flatbuffers::Offset<void> frameOrFrameDiff = 0,
+    flatbuffers::Offset<flatbuffers::Vector<int32_t>> deaths = 0,
+    int32_t frame_from_bwapi = 0,
+    int32_t battle_frame_count = 0,
+    flatbuffers::Offset<flatbuffers::Vector<int8_t>> commands_status = 0,
+    flatbuffers::Offset<flatbuffers::String> img_mode = 0,
+    const Vec2 *screen_position = 0,
+    flatbuffers::Offset<flatbuffers::Vector<uint8_t>> visibility = 0,
+    const Vec2 *visibility_size = 0,
+    flatbuffers::Offset<flatbuffers::Vector<uint8_t>> img_data = 0,
+    const Vec2 *img_size = 0) {
+  FrameStateBuilder builder_(_fbb);
+  builder_.add_img_size(img_size);
+  builder_.add_img_data(img_data);
+  builder_.add_visibility_size(visibility_size);
+  builder_.add_visibility(visibility);
+  builder_.add_screen_position(screen_position);
+  builder_.add_img_mode(img_mode);
+  builder_.add_commands_status(commands_status);
+  builder_.add_battle_frame_count(battle_frame_count);
+  builder_.add_frame_from_bwapi(frame_from_bwapi);
+  builder_.add_deaths(deaths);
+  builder_.add_frameOrFrameDiff(frameOrFrameDiff);
+  builder_.add_frameOrFrameDiff_type(frameOrFrameDiff_type);
+  return builder_.Finish();
+}
+
+inline flatbuffers::Offset<FrameState> CreateFrameStateDirect(
+    flatbuffers::FlatBufferBuilder &_fbb,
+    FrameOrFrameDiff frameOrFrameDiff_type = FrameOrFrameDiff::NONE,
+    flatbuffers::Offset<void> frameOrFrameDiff = 0,
+    const std::vector<int32_t> *deaths = nullptr,
+    int32_t frame_from_bwapi = 0,
+    int32_t battle_frame_count = 0,
+    const std::vector<int8_t> *commands_status = nullptr,
+    const char *img_mode = nullptr,
+    const Vec2 *screen_position = 0,
+    const std::vector<uint8_t> *visibility = nullptr,
+    const Vec2 *visibility_size = 0,
+    const std::vector<uint8_t> *img_data = nullptr,
+    const Vec2 *img_size = 0) {
+  return torchcraft::fbs::CreateFrameState(
+      _fbb,
+      frameOrFrameDiff_type,
+      frameOrFrameDiff,
+      deaths ? _fbb.CreateVector<int32_t>(*deaths) : 0,
+      frame_from_bwapi,
+      battle_frame_count,
+      commands_status ? _fbb.CreateVector<int8_t>(*commands_status) : 0,
+      img_mode ? _fbb.CreateString(img_mode) : 0,
+      screen_position,
+      visibility ? _fbb.CreateVector<uint8_t>(*visibility) : 0,
+      visibility_size,
+      img_data ? _fbb.CreateVector<uint8_t>(*img_data) : 0,
+      img_size);
+}
+
+flatbuffers::Offset<FrameState> CreateFrameState(flatbuffers::FlatBufferBuilder &_fbb, const FrameStateT *_o, const flatbuffers::rehasher_function_t *_rehasher = nullptr);
+
+struct PlayerLeftT : public flatbuffers::NativeTable {
+  typedef PlayerLeft TableType;
+  std::string player_left;
+  PlayerLeftT() {
+  }
+};
+
+struct PlayerLeft FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {
+  typedef PlayerLeftT NativeTableType;
+  enum {
+    VT_PLAYER_LEFT = 4
+  };
+  const flatbuffers::String *player_left() const {
+    return GetPointer<const flatbuffers::String *>(VT_PLAYER_LEFT);
+  }
+  flatbuffers::String *mutable_player_left() {
+    return GetPointer<flatbuffers::String *>(VT_PLAYER_LEFT);
+  }
+  bool Verify(flatbuffers::Verifier &verifier) const {
+    return VerifyTableStart(verifier) &&
+           VerifyOffset(verifier, VT_PLAYER_LEFT) &&
+           verifier.Verify(player_left()) &&
+           verifier.EndTable();
+  }
+  PlayerLeftT *UnPack(const flatbuffers::resolver_function_t *_resolver = nullptr) const;
+  void UnPackTo(PlayerLeftT *_o, const flatbuffers::resolver_function_t *_resolver = nullptr) const;
+  static flatbuffers::Offset<PlayerLeft> Pack(flatbuffers::FlatBufferBuilder &_fbb, const PlayerLeftT* _o, const flatbuffers::rehasher_function_t *_rehasher = nullptr);
+};
+
+struct PlayerLeftBuilder {
+  flatbuffers::FlatBufferBuilder &fbb_;
+  flatbuffers::uoffset_t start_;
+  void add_player_left(flatbuffers::Offset<flatbuffers::String> player_left) {
+    fbb_.AddOffset(PlayerLeft::VT_PLAYER_LEFT, player_left);
+  }
+  explicit PlayerLeftBuilder(flatbuffers::FlatBufferBuilder &_fbb)
+        : fbb_(_fbb) {
+    start_ = fbb_.StartTable();
+  }
+  PlayerLeftBuilder &operator=(const PlayerLeftBuilder &);
+  flatbuffers::Offset<PlayerLeft> Finish() {
+    const auto end = fbb_.EndTable(start_);
+    auto o = flatbuffers::Offset<PlayerLeft>(end);
+    return o;
+  }
+};
+
+inline flatbuffers::Offset<PlayerLeft> CreatePlayerLeft(
+    flatbuffers::FlatBufferBuilder &_fbb,
+    flatbuffers::Offset<flatbuffers::String> player_left = 0) {
+  PlayerLeftBuilder builder_(_fbb);
+  builder_.add_player_left(player_left);
+  return builder_.Finish();
+}
+
+inline flatbuffers::Offset<PlayerLeft> CreatePlayerLeftDirect(
+    flatbuffers::FlatBufferBuilder &_fbb,
+    const char *player_left = nullptr) {
+  return torchcraft::fbs::CreatePlayerLeft(
+      _fbb,
+      player_left ? _fbb.CreateString(player_left) : 0);
+}
+
+flatbuffers::Offset<PlayerLeft> CreatePlayerLeft(flatbuffers::FlatBufferBuilder &_fbb, const PlayerLeftT *_o, const flatbuffers::rehasher_function_t *_rehasher = nullptr);
+
+struct EndGameT : public flatbuffers::NativeTable {
+  typedef EndGame TableType;
+  FrameOrFrameDiffUnion frameOrFrameDiff;
+  bool game_won;
+  EndGameT()
+      : game_won(false) {
+  }
+};
+
+struct EndGame FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {
+  typedef EndGameT NativeTableType;
+  enum {
+    VT_FRAMEORFRAMEDIFF_TYPE = 4,
+    VT_FRAMEORFRAMEDIFF = 6,
+    VT_GAME_WON = 8
+  };
+  FrameOrFrameDiff frameOrFrameDiff_type() const {
+    return static_cast<FrameOrFrameDiff>(GetField<uint8_t>(VT_FRAMEORFRAMEDIFF_TYPE, 0));
+  }
+  bool mutate_frameOrFrameDiff_type(FrameOrFrameDiff _frameOrFrameDiff_type) {
+    return SetField<uint8_t>(VT_FRAMEORFRAMEDIFF_TYPE, static_cast<uint8_t>(_frameOrFrameDiff_type), 0);
+  }
+  const void *frameOrFrameDiff() const {
+    return GetPointer<const void *>(VT_FRAMEORFRAMEDIFF);
+  }
+  template<typename T> const T *frameOrFrameDiff_as() const;
+  const Frame *frameOrFrameDiff_as_Frame() const {
+    return frameOrFrameDiff_type() == FrameOrFrameDiff::Frame ? static_cast<const Frame *>(frameOrFrameDiff()) : nullptr;
+  }
+  const FrameDiff *frameOrFrameDiff_as_FrameDiff() const {
+    return frameOrFrameDiff_type() == FrameOrFrameDiff::FrameDiff ? static_cast<const FrameDiff *>(frameOrFrameDiff()) : nullptr;
+  }
+  void *mutable_frameOrFrameDiff() {
+    return GetPointer<void *>(VT_FRAMEORFRAMEDIFF);
+  }
+  bool game_won() const {
+    return GetField<uint8_t>(VT_GAME_WON, 0) != 0;
+  }
+  bool mutate_game_won(bool _game_won) {
+    return SetField<uint8_t>(VT_GAME_WON, static_cast<uint8_t>(_game_won), 0);
+  }
+  bool Verify(flatbuffers::Verifier &verifier) const {
+    return VerifyTableStart(verifier) &&
+           VerifyField<uint8_t>(verifier, VT_FRAMEORFRAMEDIFF_TYPE) &&
+           VerifyOffset(verifier, VT_FRAMEORFRAMEDIFF) &&
+           VerifyFrameOrFrameDiff(verifier, frameOrFrameDiff(), frameOrFrameDiff_type()) &&
+           VerifyField<uint8_t>(verifier, VT_GAME_WON) &&
+           verifier.EndTable();
+  }
+  EndGameT *UnPack(const flatbuffers::resolver_function_t *_resolver = nullptr) const;
+  void UnPackTo(EndGameT *_o, const flatbuffers::resolver_function_t *_resolver = nullptr) const;
+  static flatbuffers::Offset<EndGame> Pack(flatbuffers::FlatBufferBuilder &_fbb, const EndGameT* _o, const flatbuffers::rehasher_function_t *_rehasher = nullptr);
+};
+
+template<> inline const Frame *EndGame::frameOrFrameDiff_as<Frame>() const {
+  return frameOrFrameDiff_as_Frame();
+}
+
+template<> inline const FrameDiff *EndGame::frameOrFrameDiff_as<FrameDiff>() const {
+  return frameOrFrameDiff_as_FrameDiff();
+}
+
+struct EndGameBuilder {
+  flatbuffers::FlatBufferBuilder &fbb_;
+  flatbuffers::uoffset_t start_;
+  void add_frameOrFrameDiff_type(FrameOrFrameDiff frameOrFrameDiff_type) {
+    fbb_.AddElement<uint8_t>(EndGame::VT_FRAMEORFRAMEDIFF_TYPE, static_cast<uint8_t>(frameOrFrameDiff_type), 0);
+  }
+  void add_frameOrFrameDiff(flatbuffers::Offset<void> frameOrFrameDiff) {
+    fbb_.AddOffset(EndGame::VT_FRAMEORFRAMEDIFF, frameOrFrameDiff);
+  }
+  void add_game_won(bool game_won) {
+    fbb_.AddElement<uint8_t>(EndGame::VT_GAME_WON, static_cast<uint8_t>(game_won), 0);
+  }
+  explicit EndGameBuilder(flatbuffers::FlatBufferBuilder &_fbb)
+        : fbb_(_fbb) {
+    start_ = fbb_.StartTable();
+  }
+  EndGameBuilder &operator=(const EndGameBuilder &);
+  flatbuffers::Offset<EndGame> Finish() {
+    const auto end = fbb_.EndTable(start_);
+    auto o = flatbuffers::Offset<EndGame>(end);
+    return o;
+  }
+};
+
+inline flatbuffers::Offset<EndGame> CreateEndGame(
+    flatbuffers::FlatBufferBuilder &_fbb,
+    FrameOrFrameDiff frameOrFrameDiff_type = FrameOrFrameDiff::NONE,
+    flatbuffers::Offset<void> frameOrFrameDiff = 0,
+    bool game_won = false) {
+  EndGameBuilder builder_(_fbb);
+  builder_.add_frameOrFrameDiff(frameOrFrameDiff);
+  builder_.add_game_won(game_won);
+  builder_.add_frameOrFrameDiff_type(frameOrFrameDiff_type);
+  return builder_.Finish();
+}
+
+flatbuffers::Offset<EndGame> CreateEndGame(flatbuffers::FlatBufferBuilder &_fbb, const EndGameT *_o, const flatbuffers::rehasher_function_t *_rehasher = nullptr);
+
+struct ErrorT : public flatbuffers::NativeTable {
+  typedef Error TableType;
+  std::string message;
+  ErrorT() {
+  }
+};
+
+struct Error FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {
+  typedef ErrorT NativeTableType;
+  enum {
+    VT_MESSAGE = 4
+  };
+  const flatbuffers::String *message() const {
+    return GetPointer<const flatbuffers::String *>(VT_MESSAGE);
+  }
+  flatbuffers::String *mutable_message() {
+    return GetPointer<flatbuffers::String *>(VT_MESSAGE);
+  }
+  bool Verify(flatbuffers::Verifier &verifier) const {
+    return VerifyTableStart(verifier) &&
+           VerifyOffset(verifier, VT_MESSAGE) &&
+           verifier.Verify(message()) &&
+           verifier.EndTable();
+  }
+  ErrorT *UnPack(const flatbuffers::resolver_function_t *_resolver = nullptr) const;
+  void UnPackTo(ErrorT *_o, const flatbuffers::resolver_function_t *_resolver = nullptr) const;
+  static flatbuffers::Offset<Error> Pack(flatbuffers::FlatBufferBuilder &_fbb, const ErrorT* _o, const flatbuffers::rehasher_function_t *_rehasher = nullptr);
+};
+
+struct ErrorBuilder {
+  flatbuffers::FlatBufferBuilder &fbb_;
+  flatbuffers::uoffset_t start_;
+  void add_message(flatbuffers::Offset<flatbuffers::String> message) {
+    fbb_.AddOffset(Error::VT_MESSAGE, message);
+  }
+  explicit ErrorBuilder(flatbuffers::FlatBufferBuilder &_fbb)
+        : fbb_(_fbb) {
+    start_ = fbb_.StartTable();
+  }
+  ErrorBuilder &operator=(const ErrorBuilder &);
+  flatbuffers::Offset<Error> Finish() {
+    const auto end = fbb_.EndTable(start_);
+    auto o = flatbuffers::Offset<Error>(end);
+    return o;
+  }
+};
+
+inline flatbuffers::Offset<Error> CreateError(
+    flatbuffers::FlatBufferBuilder &_fbb,
+    flatbuffers::Offset<flatbuffers::String> message = 0) {
+  ErrorBuilder builder_(_fbb);
+  builder_.add_message(message);
+  return builder_.Finish();
+}
+
+inline flatbuffers::Offset<Error> CreateErrorDirect(
+    flatbuffers::FlatBufferBuilder &_fbb,
+    const char *message = nullptr) {
+  return torchcraft::fbs::CreateError(
+      _fbb,
+      message ? _fbb.CreateString(message) : 0);
+}
+
+flatbuffers::Offset<Error> CreateError(flatbuffers::FlatBufferBuilder &_fbb, const ErrorT *_o, const flatbuffers::rehasher_function_t *_rehasher = nullptr);
+
+struct MessageT : public flatbuffers::NativeTable {
+  typedef Message TableType;
+  MessageTypeUnion msg;
+  std::string uid;
+  MessageT() {
+  }
+};
+
+struct Message FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {
+  typedef MessageT NativeTableType;
+  enum {
+    VT_MSG_TYPE = 4,
+    VT_MSG = 6,
+    VT_UID = 8
+  };
+  MessageType msg_type() const {
+    return static_cast<MessageType>(GetField<uint8_t>(VT_MSG_TYPE, 0));
+  }
+  bool mutate_msg_type(MessageType _msg_type) {
+    return SetField<uint8_t>(VT_MSG_TYPE, static_cast<uint8_t>(_msg_type), 0);
+  }
+  const void *msg() const {
+    return GetPointer<const void *>(VT_MSG);
+  }
+  template<typename T> const T *msg_as() const;
+  const HandshakeClient *msg_as_HandshakeClient() const {
+    return msg_type() == MessageType::HandshakeClient ? static_cast<const HandshakeClient *>(msg()) : nullptr;
+  }
+  const Commands *msg_as_Commands() const {
+    return msg_type() == MessageType::Commands ? static_cast<const Commands *>(msg()) : nullptr;
+  }
+  const HandshakeServer *msg_as_HandshakeServer() const {
+    return msg_type() == MessageType::HandshakeServer ? static_cast<const HandshakeServer *>(msg()) : nullptr;
+  }
+  const FrameState *msg_as_FrameState() const {
+    return msg_type() == MessageType::FrameState ? static_cast<const FrameState *>(msg()) : nullptr;
+  }
+  const PlayerLeft *msg_as_PlayerLeft() const {
+    return msg_type() == MessageType::PlayerLeft ? static_cast<const PlayerLeft *>(msg()) : nullptr;
+  }
+  const EndGame *msg_as_EndGame() const {
+    return msg_type() == MessageType::EndGame ? static_cast<const EndGame *>(msg()) : nullptr;
+  }
+  const Error *msg_as_Error() const {
+    return msg_type() == MessageType::Error ? static_cast<const Error *>(msg()) : nullptr;
+  }
+  void *mutable_msg() {
+    return GetPointer<void *>(VT_MSG);
+  }
+  const flatbuffers::String *uid() const {
+    return GetPointer<const flatbuffers::String *>(VT_UID);
+  }
+  flatbuffers::String *mutable_uid() {
+    return GetPointer<flatbuffers::String *>(VT_UID);
+  }
+  bool Verify(flatbuffers::Verifier &verifier) const {
+    return VerifyTableStart(verifier) &&
+           VerifyField<uint8_t>(verifier, VT_MSG_TYPE) &&
+           VerifyOffset(verifier, VT_MSG) &&
+           VerifyMessageType(verifier, msg(), msg_type()) &&
+           VerifyOffset(verifier, VT_UID) &&
+           verifier.Verify(uid()) &&
+           verifier.EndTable();
+  }
+  MessageT *UnPack(const flatbuffers::resolver_function_t *_resolver = nullptr) const;
+  void UnPackTo(MessageT *_o, const flatbuffers::resolver_function_t *_resolver = nullptr) const;
+  static flatbuffers::Offset<Message> Pack(flatbuffers::FlatBufferBuilder &_fbb, const MessageT* _o, const flatbuffers::rehasher_function_t *_rehasher = nullptr);
+};
+
+template<> inline const HandshakeClient *Message::msg_as<HandshakeClient>() const {
+  return msg_as_HandshakeClient();
+}
+
+template<> inline const Commands *Message::msg_as<Commands>() const {
+  return msg_as_Commands();
+}
+
+template<> inline const HandshakeServer *Message::msg_as<HandshakeServer>() const {
+  return msg_as_HandshakeServer();
+}
+
+template<> inline const FrameState *Message::msg_as<FrameState>() const {
+  return msg_as_FrameState();
+}
+
+template<> inline const PlayerLeft *Message::msg_as<PlayerLeft>() const {
+  return msg_as_PlayerLeft();
+}
+
+template<> inline const EndGame *Message::msg_as<EndGame>() const {
+  return msg_as_EndGame();
+}
+
+template<> inline const Error *Message::msg_as<Error>() const {
+  return msg_as_Error();
+}
+
+struct MessageBuilder {
+  flatbuffers::FlatBufferBuilder &fbb_;
+  flatbuffers::uoffset_t start_;
+  void add_msg_type(MessageType msg_type) {
+    fbb_.AddElement<uint8_t>(Message::VT_MSG_TYPE, static_cast<uint8_t>(msg_type), 0);
+  }
+  void add_msg(flatbuffers::Offset<void> msg) {
+    fbb_.AddOffset(Message::VT_MSG, msg);
+  }
+  void add_uid(flatbuffers::Offset<flatbuffers::String> uid) {
+    fbb_.AddOffset(Message::VT_UID, uid);
+  }
+  explicit MessageBuilder(flatbuffers::FlatBufferBuilder &_fbb)
+        : fbb_(_fbb) {
+    start_ = fbb_.StartTable();
+  }
+  MessageBuilder &operator=(const MessageBuilder &);
+  flatbuffers::Offset<Message> Finish() {
+    const auto end = fbb_.EndTable(start_);
+    auto o = flatbuffers::Offset<Message>(end);
+    return o;
+  }
+};
+
+inline flatbuffers::Offset<Message> CreateMessage(
+    flatbuffers::FlatBufferBuilder &_fbb,
+    MessageType msg_type = MessageType::NONE,
+    flatbuffers::Offset<void> msg = 0,
+    flatbuffers::Offset<flatbuffers::String> uid = 0) {
+  MessageBuilder builder_(_fbb);
+  builder_.add_uid(uid);
+  builder_.add_msg(msg);
+  builder_.add_msg_type(msg_type);
+  return builder_.Finish();
+}
+
+inline flatbuffers::Offset<Message> CreateMessageDirect(
+    flatbuffers::FlatBufferBuilder &_fbb,
+    MessageType msg_type = MessageType::NONE,
+    flatbuffers::Offset<void> msg = 0,
+    const char *uid = nullptr) {
+  return torchcraft::fbs::CreateMessage(
+      _fbb,
+      msg_type,
+      msg,
+      uid ? _fbb.CreateString(uid) : 0);
+}
+
+flatbuffers::Offset<Message> CreateMessage(flatbuffers::FlatBufferBuilder &_fbb, const MessageT *_o, const flatbuffers::rehasher_function_t *_rehasher = nullptr);
 
 struct MapT : public flatbuffers::NativeTable {
   typedef Map TableType;
@@ -2553,6 +4215,371 @@ inline flatbuffers::Offset<Replayer> CreateReplayerDirect(
 
 flatbuffers::Offset<Replayer> CreateReplayer(flatbuffers::FlatBufferBuilder &_fbb, const ReplayerT *_o, const flatbuffers::rehasher_function_t *_rehasher = nullptr);
 
+inline CommandT *Command::UnPack(const flatbuffers::resolver_function_t *_resolver) const {
+  auto _o = new CommandT();
+  UnPackTo(_o, _resolver);
+  return _o;
+}
+
+inline void Command::UnPackTo(CommandT *_o, const flatbuffers::resolver_function_t *_resolver) const {
+  (void)_o;
+  (void)_resolver;
+  { auto _e = code(); _o->code = _e; };
+  { auto _e = args(); if (_e) { _o->args.resize(_e->size()); for (flatbuffers::uoffset_t _i = 0; _i < _e->size(); _i++) { _o->args[_i] = _e->Get(_i); } } };
+  { auto _e = str(); if (_e) _o->str = _e->str(); };
+}
+
+inline flatbuffers::Offset<Command> Command::Pack(flatbuffers::FlatBufferBuilder &_fbb, const CommandT* _o, const flatbuffers::rehasher_function_t *_rehasher) {
+  return CreateCommand(_fbb, _o, _rehasher);
+}
+
+inline flatbuffers::Offset<Command> CreateCommand(flatbuffers::FlatBufferBuilder &_fbb, const CommandT *_o, const flatbuffers::rehasher_function_t *_rehasher) {
+  (void)_rehasher;
+  (void)_o;
+  struct _VectorArgs { flatbuffers::FlatBufferBuilder *__fbb; const CommandT* __o; const flatbuffers::rehasher_function_t *__rehasher; } _va = { &_fbb, _o, _rehasher}; (void)_va;
+  auto _code = _o->code;
+  auto _args = _o->args.size() ? _fbb.CreateVector(_o->args) : 0;
+  auto _str = _o->str.empty() ? 0 : _fbb.CreateString(_o->str);
+  return torchcraft::fbs::CreateCommand(
+      _fbb,
+      _code,
+      _args,
+      _str);
+}
+
+inline PlayerT *Player::UnPack(const flatbuffers::resolver_function_t *_resolver) const {
+  auto _o = new PlayerT();
+  UnPackTo(_o, _resolver);
+  return _o;
+}
+
+inline void Player::UnPackTo(PlayerT *_o, const flatbuffers::resolver_function_t *_resolver) const {
+  (void)_o;
+  (void)_resolver;
+  { auto _e = id(); _o->id = _e; };
+  { auto _e = race(); _o->race = _e; };
+  { auto _e = name(); if (_e) _o->name = _e->str(); };
+  { auto _e = is_enemy(); _o->is_enemy = _e; };
+}
+
+inline flatbuffers::Offset<Player> Player::Pack(flatbuffers::FlatBufferBuilder &_fbb, const PlayerT* _o, const flatbuffers::rehasher_function_t *_rehasher) {
+  return CreatePlayer(_fbb, _o, _rehasher);
+}
+
+inline flatbuffers::Offset<Player> CreatePlayer(flatbuffers::FlatBufferBuilder &_fbb, const PlayerT *_o, const flatbuffers::rehasher_function_t *_rehasher) {
+  (void)_rehasher;
+  (void)_o;
+  struct _VectorArgs { flatbuffers::FlatBufferBuilder *__fbb; const PlayerT* __o; const flatbuffers::rehasher_function_t *__rehasher; } _va = { &_fbb, _o, _rehasher}; (void)_va;
+  auto _id = _o->id;
+  auto _race = _o->race;
+  auto _name = _o->name.empty() ? 0 : _fbb.CreateString(_o->name);
+  auto _is_enemy = _o->is_enemy;
+  return torchcraft::fbs::CreatePlayer(
+      _fbb,
+      _id,
+      _race,
+      _name,
+      _is_enemy);
+}
+
+inline HandshakeClientT *HandshakeClient::UnPack(const flatbuffers::resolver_function_t *_resolver) const {
+  auto _o = new HandshakeClientT();
+  UnPackTo(_o, _resolver);
+  return _o;
+}
+
+inline void HandshakeClient::UnPackTo(HandshakeClientT *_o, const flatbuffers::resolver_function_t *_resolver) const {
+  (void)_o;
+  (void)_resolver;
+  { auto _e = protocol(); _o->protocol = _e; };
+  { auto _e = map(); if (_e) _o->map = _e->str(); };
+  { auto _e = window_size(); if (_e) _o->window_size = std::unique_ptr<Vec2>(new Vec2(*_e)); };
+  { auto _e = window_pos(); if (_e) _o->window_pos = std::unique_ptr<Vec2>(new Vec2(*_e)); };
+  { auto _e = micro_mode(); _o->micro_mode = _e; };
+}
+
+inline flatbuffers::Offset<HandshakeClient> HandshakeClient::Pack(flatbuffers::FlatBufferBuilder &_fbb, const HandshakeClientT* _o, const flatbuffers::rehasher_function_t *_rehasher) {
+  return CreateHandshakeClient(_fbb, _o, _rehasher);
+}
+
+inline flatbuffers::Offset<HandshakeClient> CreateHandshakeClient(flatbuffers::FlatBufferBuilder &_fbb, const HandshakeClientT *_o, const flatbuffers::rehasher_function_t *_rehasher) {
+  (void)_rehasher;
+  (void)_o;
+  struct _VectorArgs { flatbuffers::FlatBufferBuilder *__fbb; const HandshakeClientT* __o; const flatbuffers::rehasher_function_t *__rehasher; } _va = { &_fbb, _o, _rehasher}; (void)_va;
+  auto _protocol = _o->protocol;
+  auto _map = _o->map.empty() ? 0 : _fbb.CreateString(_o->map);
+  auto _window_size = _o->window_size ? _o->window_size.get() : 0;
+  auto _window_pos = _o->window_pos ? _o->window_pos.get() : 0;
+  auto _micro_mode = _o->micro_mode;
+  return torchcraft::fbs::CreateHandshakeClient(
+      _fbb,
+      _protocol,
+      _map,
+      _window_size,
+      _window_pos,
+      _micro_mode);
+}
+
+inline HandshakeServerT *HandshakeServer::UnPack(const flatbuffers::resolver_function_t *_resolver) const {
+  auto _o = new HandshakeServerT();
+  UnPackTo(_o, _resolver);
+  return _o;
+}
+
+inline void HandshakeServer::UnPackTo(HandshakeServerT *_o, const flatbuffers::resolver_function_t *_resolver) const {
+  (void)_o;
+  (void)_resolver;
+  { auto _e = lag_frames(); _o->lag_frames = _e; };
+  { auto _e = map_size(); if (_e) _o->map_size = std::unique_ptr<Vec2>(new Vec2(*_e)); };
+  { auto _e = ground_height_data(); if (_e) { _o->ground_height_data.resize(_e->size()); for (flatbuffers::uoffset_t _i = 0; _i < _e->size(); _i++) { _o->ground_height_data[_i] = _e->Get(_i); } } };
+  { auto _e = walkable_data(); if (_e) { _o->walkable_data.resize(_e->size()); for (flatbuffers::uoffset_t _i = 0; _i < _e->size(); _i++) { _o->walkable_data[_i] = _e->Get(_i); } } };
+  { auto _e = map_name(); if (_e) _o->map_name = _e->str(); };
+  { auto _e = is_replay(); _o->is_replay = _e; };
+  { auto _e = player_id(); _o->player_id = _e; };
+  { auto _e = neutral_id(); _o->neutral_id = _e; };
+  { auto _e = battle_frame_count(); _o->battle_frame_count = _e; };
+  { auto _e = buildable_data(); if (_e) { _o->buildable_data.resize(_e->size()); for (flatbuffers::uoffset_t _i = 0; _i < _e->size(); _i++) { _o->buildable_data[_i] = _e->Get(_i); } } };
+  { auto _e = start_locations(); if (_e) { _o->start_locations.resize(_e->size()); for (flatbuffers::uoffset_t _i = 0; _i < _e->size(); _i++) { _o->start_locations[_i] = *_e->Get(_i); } } };
+  { auto _e = players(); if (_e) { _o->players.resize(_e->size()); for (flatbuffers::uoffset_t _i = 0; _i < _e->size(); _i++) { _o->players[_i] = std::unique_ptr<PlayerT>(_e->Get(_i)->UnPack(_resolver)); } } };
+}
+
+inline flatbuffers::Offset<HandshakeServer> HandshakeServer::Pack(flatbuffers::FlatBufferBuilder &_fbb, const HandshakeServerT* _o, const flatbuffers::rehasher_function_t *_rehasher) {
+  return CreateHandshakeServer(_fbb, _o, _rehasher);
+}
+
+inline flatbuffers::Offset<HandshakeServer> CreateHandshakeServer(flatbuffers::FlatBufferBuilder &_fbb, const HandshakeServerT *_o, const flatbuffers::rehasher_function_t *_rehasher) {
+  (void)_rehasher;
+  (void)_o;
+  struct _VectorArgs { flatbuffers::FlatBufferBuilder *__fbb; const HandshakeServerT* __o; const flatbuffers::rehasher_function_t *__rehasher; } _va = { &_fbb, _o, _rehasher}; (void)_va;
+  auto _lag_frames = _o->lag_frames;
+  auto _map_size = _o->map_size ? _o->map_size.get() : 0;
+  auto _ground_height_data = _o->ground_height_data.size() ? _fbb.CreateVector(_o->ground_height_data) : 0;
+  auto _walkable_data = _o->walkable_data.size() ? _fbb.CreateVector(_o->walkable_data) : 0;
+  auto _map_name = _o->map_name.empty() ? 0 : _fbb.CreateString(_o->map_name);
+  auto _is_replay = _o->is_replay;
+  auto _player_id = _o->player_id;
+  auto _neutral_id = _o->neutral_id;
+  auto _battle_frame_count = _o->battle_frame_count;
+  auto _buildable_data = _o->buildable_data.size() ? _fbb.CreateVector(_o->buildable_data) : 0;
+  auto _start_locations = _o->start_locations.size() ? _fbb.CreateVectorOfStructs(_o->start_locations) : 0;
+  auto _players = _o->players.size() ? _fbb.CreateVector<flatbuffers::Offset<Player>> (_o->players.size(), [](size_t i, _VectorArgs *__va) { return CreatePlayer(*__va->__fbb, __va->__o->players[i].get(), __va->__rehasher); }, &_va ) : 0;
+  return torchcraft::fbs::CreateHandshakeServer(
+      _fbb,
+      _lag_frames,
+      _map_size,
+      _ground_height_data,
+      _walkable_data,
+      _map_name,
+      _is_replay,
+      _player_id,
+      _neutral_id,
+      _battle_frame_count,
+      _buildable_data,
+      _start_locations,
+      _players);
+}
+
+inline CommandsT *Commands::UnPack(const flatbuffers::resolver_function_t *_resolver) const {
+  auto _o = new CommandsT();
+  UnPackTo(_o, _resolver);
+  return _o;
+}
+
+inline void Commands::UnPackTo(CommandsT *_o, const flatbuffers::resolver_function_t *_resolver) const {
+  (void)_o;
+  (void)_resolver;
+  { auto _e = commands(); if (_e) { _o->commands.resize(_e->size()); for (flatbuffers::uoffset_t _i = 0; _i < _e->size(); _i++) { _o->commands[_i] = std::unique_ptr<CommandT>(_e->Get(_i)->UnPack(_resolver)); } } };
+}
+
+inline flatbuffers::Offset<Commands> Commands::Pack(flatbuffers::FlatBufferBuilder &_fbb, const CommandsT* _o, const flatbuffers::rehasher_function_t *_rehasher) {
+  return CreateCommands(_fbb, _o, _rehasher);
+}
+
+inline flatbuffers::Offset<Commands> CreateCommands(flatbuffers::FlatBufferBuilder &_fbb, const CommandsT *_o, const flatbuffers::rehasher_function_t *_rehasher) {
+  (void)_rehasher;
+  (void)_o;
+  struct _VectorArgs { flatbuffers::FlatBufferBuilder *__fbb; const CommandsT* __o; const flatbuffers::rehasher_function_t *__rehasher; } _va = { &_fbb, _o, _rehasher}; (void)_va;
+  auto _commands = _o->commands.size() ? _fbb.CreateVector<flatbuffers::Offset<Command>> (_o->commands.size(), [](size_t i, _VectorArgs *__va) { return CreateCommand(*__va->__fbb, __va->__o->commands[i].get(), __va->__rehasher); }, &_va ) : 0;
+  return torchcraft::fbs::CreateCommands(
+      _fbb,
+      _commands);
+}
+
+inline FrameStateT *FrameState::UnPack(const flatbuffers::resolver_function_t *_resolver) const {
+  auto _o = new FrameStateT();
+  UnPackTo(_o, _resolver);
+  return _o;
+}
+
+inline void FrameState::UnPackTo(FrameStateT *_o, const flatbuffers::resolver_function_t *_resolver) const {
+  (void)_o;
+  (void)_resolver;
+  { auto _e = frameOrFrameDiff_type(); _o->frameOrFrameDiff.type = _e; };
+  { auto _e = frameOrFrameDiff(); if (_e) _o->frameOrFrameDiff.value = FrameOrFrameDiffUnion::UnPack(_e, frameOrFrameDiff_type(), _resolver); };
+  { auto _e = deaths(); if (_e) { _o->deaths.resize(_e->size()); for (flatbuffers::uoffset_t _i = 0; _i < _e->size(); _i++) { _o->deaths[_i] = _e->Get(_i); } } };
+  { auto _e = frame_from_bwapi(); _o->frame_from_bwapi = _e; };
+  { auto _e = battle_frame_count(); _o->battle_frame_count = _e; };
+  { auto _e = commands_status(); if (_e) { _o->commands_status.resize(_e->size()); for (flatbuffers::uoffset_t _i = 0; _i < _e->size(); _i++) { _o->commands_status[_i] = _e->Get(_i); } } };
+  { auto _e = img_mode(); if (_e) _o->img_mode = _e->str(); };
+  { auto _e = screen_position(); if (_e) _o->screen_position = std::unique_ptr<Vec2>(new Vec2(*_e)); };
+  { auto _e = visibility(); if (_e) { _o->visibility.resize(_e->size()); for (flatbuffers::uoffset_t _i = 0; _i < _e->size(); _i++) { _o->visibility[_i] = _e->Get(_i); } } };
+  { auto _e = visibility_size(); if (_e) _o->visibility_size = std::unique_ptr<Vec2>(new Vec2(*_e)); };
+  { auto _e = img_data(); if (_e) { _o->img_data.resize(_e->size()); for (flatbuffers::uoffset_t _i = 0; _i < _e->size(); _i++) { _o->img_data[_i] = _e->Get(_i); } } };
+  { auto _e = img_size(); if (_e) _o->img_size = std::unique_ptr<Vec2>(new Vec2(*_e)); };
+}
+
+inline flatbuffers::Offset<FrameState> FrameState::Pack(flatbuffers::FlatBufferBuilder &_fbb, const FrameStateT* _o, const flatbuffers::rehasher_function_t *_rehasher) {
+  return CreateFrameState(_fbb, _o, _rehasher);
+}
+
+inline flatbuffers::Offset<FrameState> CreateFrameState(flatbuffers::FlatBufferBuilder &_fbb, const FrameStateT *_o, const flatbuffers::rehasher_function_t *_rehasher) {
+  (void)_rehasher;
+  (void)_o;
+  struct _VectorArgs { flatbuffers::FlatBufferBuilder *__fbb; const FrameStateT* __o; const flatbuffers::rehasher_function_t *__rehasher; } _va = { &_fbb, _o, _rehasher}; (void)_va;
+  auto _frameOrFrameDiff_type = _o->frameOrFrameDiff.type;
+  auto _frameOrFrameDiff = _o->frameOrFrameDiff.Pack(_fbb);
+  auto _deaths = _o->deaths.size() ? _fbb.CreateVector(_o->deaths) : 0;
+  auto _frame_from_bwapi = _o->frame_from_bwapi;
+  auto _battle_frame_count = _o->battle_frame_count;
+  auto _commands_status = _o->commands_status.size() ? _fbb.CreateVector(_o->commands_status) : 0;
+  auto _img_mode = _o->img_mode.empty() ? 0 : _fbb.CreateString(_o->img_mode);
+  auto _screen_position = _o->screen_position ? _o->screen_position.get() : 0;
+  auto _visibility = _o->visibility.size() ? _fbb.CreateVector(_o->visibility) : 0;
+  auto _visibility_size = _o->visibility_size ? _o->visibility_size.get() : 0;
+  auto _img_data = _o->img_data.size() ? _fbb.CreateVector(_o->img_data) : 0;
+  auto _img_size = _o->img_size ? _o->img_size.get() : 0;
+  return torchcraft::fbs::CreateFrameState(
+      _fbb,
+      _frameOrFrameDiff_type,
+      _frameOrFrameDiff,
+      _deaths,
+      _frame_from_bwapi,
+      _battle_frame_count,
+      _commands_status,
+      _img_mode,
+      _screen_position,
+      _visibility,
+      _visibility_size,
+      _img_data,
+      _img_size);
+}
+
+inline PlayerLeftT *PlayerLeft::UnPack(const flatbuffers::resolver_function_t *_resolver) const {
+  auto _o = new PlayerLeftT();
+  UnPackTo(_o, _resolver);
+  return _o;
+}
+
+inline void PlayerLeft::UnPackTo(PlayerLeftT *_o, const flatbuffers::resolver_function_t *_resolver) const {
+  (void)_o;
+  (void)_resolver;
+  { auto _e = player_left(); if (_e) _o->player_left = _e->str(); };
+}
+
+inline flatbuffers::Offset<PlayerLeft> PlayerLeft::Pack(flatbuffers::FlatBufferBuilder &_fbb, const PlayerLeftT* _o, const flatbuffers::rehasher_function_t *_rehasher) {
+  return CreatePlayerLeft(_fbb, _o, _rehasher);
+}
+
+inline flatbuffers::Offset<PlayerLeft> CreatePlayerLeft(flatbuffers::FlatBufferBuilder &_fbb, const PlayerLeftT *_o, const flatbuffers::rehasher_function_t *_rehasher) {
+  (void)_rehasher;
+  (void)_o;
+  struct _VectorArgs { flatbuffers::FlatBufferBuilder *__fbb; const PlayerLeftT* __o; const flatbuffers::rehasher_function_t *__rehasher; } _va = { &_fbb, _o, _rehasher}; (void)_va;
+  auto _player_left = _o->player_left.empty() ? 0 : _fbb.CreateString(_o->player_left);
+  return torchcraft::fbs::CreatePlayerLeft(
+      _fbb,
+      _player_left);
+}
+
+inline EndGameT *EndGame::UnPack(const flatbuffers::resolver_function_t *_resolver) const {
+  auto _o = new EndGameT();
+  UnPackTo(_o, _resolver);
+  return _o;
+}
+
+inline void EndGame::UnPackTo(EndGameT *_o, const flatbuffers::resolver_function_t *_resolver) const {
+  (void)_o;
+  (void)_resolver;
+  { auto _e = frameOrFrameDiff_type(); _o->frameOrFrameDiff.type = _e; };
+  { auto _e = frameOrFrameDiff(); if (_e) _o->frameOrFrameDiff.value = FrameOrFrameDiffUnion::UnPack(_e, frameOrFrameDiff_type(), _resolver); };
+  { auto _e = game_won(); _o->game_won = _e; };
+}
+
+inline flatbuffers::Offset<EndGame> EndGame::Pack(flatbuffers::FlatBufferBuilder &_fbb, const EndGameT* _o, const flatbuffers::rehasher_function_t *_rehasher) {
+  return CreateEndGame(_fbb, _o, _rehasher);
+}
+
+inline flatbuffers::Offset<EndGame> CreateEndGame(flatbuffers::FlatBufferBuilder &_fbb, const EndGameT *_o, const flatbuffers::rehasher_function_t *_rehasher) {
+  (void)_rehasher;
+  (void)_o;
+  struct _VectorArgs { flatbuffers::FlatBufferBuilder *__fbb; const EndGameT* __o; const flatbuffers::rehasher_function_t *__rehasher; } _va = { &_fbb, _o, _rehasher}; (void)_va;
+  auto _frameOrFrameDiff_type = _o->frameOrFrameDiff.type;
+  auto _frameOrFrameDiff = _o->frameOrFrameDiff.Pack(_fbb);
+  auto _game_won = _o->game_won;
+  return torchcraft::fbs::CreateEndGame(
+      _fbb,
+      _frameOrFrameDiff_type,
+      _frameOrFrameDiff,
+      _game_won);
+}
+
+inline ErrorT *Error::UnPack(const flatbuffers::resolver_function_t *_resolver) const {
+  auto _o = new ErrorT();
+  UnPackTo(_o, _resolver);
+  return _o;
+}
+
+inline void Error::UnPackTo(ErrorT *_o, const flatbuffers::resolver_function_t *_resolver) const {
+  (void)_o;
+  (void)_resolver;
+  { auto _e = message(); if (_e) _o->message = _e->str(); };
+}
+
+inline flatbuffers::Offset<Error> Error::Pack(flatbuffers::FlatBufferBuilder &_fbb, const ErrorT* _o, const flatbuffers::rehasher_function_t *_rehasher) {
+  return CreateError(_fbb, _o, _rehasher);
+}
+
+inline flatbuffers::Offset<Error> CreateError(flatbuffers::FlatBufferBuilder &_fbb, const ErrorT *_o, const flatbuffers::rehasher_function_t *_rehasher) {
+  (void)_rehasher;
+  (void)_o;
+  struct _VectorArgs { flatbuffers::FlatBufferBuilder *__fbb; const ErrorT* __o; const flatbuffers::rehasher_function_t *__rehasher; } _va = { &_fbb, _o, _rehasher}; (void)_va;
+  auto _message = _o->message.empty() ? 0 : _fbb.CreateString(_o->message);
+  return torchcraft::fbs::CreateError(
+      _fbb,
+      _message);
+}
+
+inline MessageT *Message::UnPack(const flatbuffers::resolver_function_t *_resolver) const {
+  auto _o = new MessageT();
+  UnPackTo(_o, _resolver);
+  return _o;
+}
+
+inline void Message::UnPackTo(MessageT *_o, const flatbuffers::resolver_function_t *_resolver) const {
+  (void)_o;
+  (void)_resolver;
+  { auto _e = msg_type(); _o->msg.type = _e; };
+  { auto _e = msg(); if (_e) _o->msg.value = MessageTypeUnion::UnPack(_e, msg_type(), _resolver); };
+  { auto _e = uid(); if (_e) _o->uid = _e->str(); };
+}
+
+inline flatbuffers::Offset<Message> Message::Pack(flatbuffers::FlatBufferBuilder &_fbb, const MessageT* _o, const flatbuffers::rehasher_function_t *_rehasher) {
+  return CreateMessage(_fbb, _o, _rehasher);
+}
+
+inline flatbuffers::Offset<Message> CreateMessage(flatbuffers::FlatBufferBuilder &_fbb, const MessageT *_o, const flatbuffers::rehasher_function_t *_rehasher) {
+  (void)_rehasher;
+  (void)_o;
+  struct _VectorArgs { flatbuffers::FlatBufferBuilder *__fbb; const MessageT* __o; const flatbuffers::rehasher_function_t *__rehasher; } _va = { &_fbb, _o, _rehasher}; (void)_va;
+  auto _msg_type = _o->msg.type;
+  auto _msg = _o->msg.Pack(_fbb);
+  auto _uid = _o->uid.empty() ? 0 : _fbb.CreateString(_o->uid);
+  return torchcraft::fbs::CreateMessage(
+      _fbb,
+      _msg_type,
+      _msg,
+      _uid);
+}
+
 inline MapT *Map::UnPack(const flatbuffers::resolver_function_t *_resolver) const {
   auto _o = new MapT();
   UnPackTo(_o, _resolver);
@@ -3233,32 +5260,315 @@ inline flatbuffers::Offset<Replayer> CreateReplayer(flatbuffers::FlatBufferBuild
       _keyFrame);
 }
 
-inline const torchcraft::fbs::Replayer *GetReplayer(const void *buf) {
-  return flatbuffers::GetRoot<torchcraft::fbs::Replayer>(buf);
+inline bool VerifyFrameOrFrameDiff(flatbuffers::Verifier &verifier, const void *obj, FrameOrFrameDiff type) {
+  switch (type) {
+    case FrameOrFrameDiff::NONE: {
+      return true;
+    }
+    case FrameOrFrameDiff::Frame: {
+      auto ptr = reinterpret_cast<const Frame *>(obj);
+      return verifier.VerifyTable(ptr);
+    }
+    case FrameOrFrameDiff::FrameDiff: {
+      auto ptr = reinterpret_cast<const FrameDiff *>(obj);
+      return verifier.VerifyTable(ptr);
+    }
+    default: return false;
+  }
 }
 
-inline Replayer *GetMutableReplayer(void *buf) {
-  return flatbuffers::GetMutableRoot<Replayer>(buf);
+inline bool VerifyFrameOrFrameDiffVector(flatbuffers::Verifier &verifier, const flatbuffers::Vector<flatbuffers::Offset<void>> *values, const flatbuffers::Vector<uint8_t> *types) {
+  if (values->size() != types->size()) return false;
+  for (flatbuffers::uoffset_t i = 0; i < values->size(); ++i) {
+    if (!VerifyFrameOrFrameDiff(
+        verifier,  values->Get(i), types->GetEnum<FrameOrFrameDiff>(i))) {
+      return false;
+    }
+  }
+  return true;
 }
 
-inline bool VerifyReplayerBuffer(
+inline void *FrameOrFrameDiffUnion::UnPack(const void *obj, FrameOrFrameDiff type, const flatbuffers::resolver_function_t *resolver) {
+  switch (type) {
+    case FrameOrFrameDiff::Frame: {
+      auto ptr = reinterpret_cast<const Frame *>(obj);
+      return ptr->UnPack(resolver);
+    }
+    case FrameOrFrameDiff::FrameDiff: {
+      auto ptr = reinterpret_cast<const FrameDiff *>(obj);
+      return ptr->UnPack(resolver);
+    }
+    default: return nullptr;
+  }
+}
+
+inline flatbuffers::Offset<void> FrameOrFrameDiffUnion::Pack(flatbuffers::FlatBufferBuilder &_fbb, const flatbuffers::rehasher_function_t *_rehasher) const {
+  switch (type) {
+    case FrameOrFrameDiff::Frame: {
+      auto ptr = reinterpret_cast<const FrameT *>(value);
+      return CreateFrame(_fbb, ptr, _rehasher).Union();
+    }
+    case FrameOrFrameDiff::FrameDiff: {
+      auto ptr = reinterpret_cast<const FrameDiffT *>(value);
+      return CreateFrameDiff(_fbb, ptr, _rehasher).Union();
+    }
+    default: return 0;
+  }
+}
+
+inline FrameOrFrameDiffUnion::FrameOrFrameDiffUnion(const FrameOrFrameDiffUnion &u) FLATBUFFERS_NOEXCEPT : type(u.type), value(nullptr) {
+  switch (type) {
+    case FrameOrFrameDiff::Frame: {
+      assert(false);  // FrameT not copyable.
+      break;
+    }
+    case FrameOrFrameDiff::FrameDiff: {
+      assert(false);  // FrameDiffT not copyable.
+      break;
+    }
+    default:
+      break;
+  }
+}
+
+inline void FrameOrFrameDiffUnion::Reset() {
+  switch (type) {
+    case FrameOrFrameDiff::Frame: {
+      auto ptr = reinterpret_cast<FrameT *>(value);
+      delete ptr;
+      break;
+    }
+    case FrameOrFrameDiff::FrameDiff: {
+      auto ptr = reinterpret_cast<FrameDiffT *>(value);
+      delete ptr;
+      break;
+    }
+    default: break;
+  }
+  value = nullptr;
+  type = FrameOrFrameDiff::NONE;
+}
+
+inline bool VerifyMessageType(flatbuffers::Verifier &verifier, const void *obj, MessageType type) {
+  switch (type) {
+    case MessageType::NONE: {
+      return true;
+    }
+    case MessageType::HandshakeClient: {
+      auto ptr = reinterpret_cast<const HandshakeClient *>(obj);
+      return verifier.VerifyTable(ptr);
+    }
+    case MessageType::Commands: {
+      auto ptr = reinterpret_cast<const Commands *>(obj);
+      return verifier.VerifyTable(ptr);
+    }
+    case MessageType::HandshakeServer: {
+      auto ptr = reinterpret_cast<const HandshakeServer *>(obj);
+      return verifier.VerifyTable(ptr);
+    }
+    case MessageType::FrameState: {
+      auto ptr = reinterpret_cast<const FrameState *>(obj);
+      return verifier.VerifyTable(ptr);
+    }
+    case MessageType::PlayerLeft: {
+      auto ptr = reinterpret_cast<const PlayerLeft *>(obj);
+      return verifier.VerifyTable(ptr);
+    }
+    case MessageType::EndGame: {
+      auto ptr = reinterpret_cast<const EndGame *>(obj);
+      return verifier.VerifyTable(ptr);
+    }
+    case MessageType::Error: {
+      auto ptr = reinterpret_cast<const Error *>(obj);
+      return verifier.VerifyTable(ptr);
+    }
+    default: return false;
+  }
+}
+
+inline bool VerifyMessageTypeVector(flatbuffers::Verifier &verifier, const flatbuffers::Vector<flatbuffers::Offset<void>> *values, const flatbuffers::Vector<uint8_t> *types) {
+  if (values->size() != types->size()) return false;
+  for (flatbuffers::uoffset_t i = 0; i < values->size(); ++i) {
+    if (!VerifyMessageType(
+        verifier,  values->Get(i), types->GetEnum<MessageType>(i))) {
+      return false;
+    }
+  }
+  return true;
+}
+
+inline void *MessageTypeUnion::UnPack(const void *obj, MessageType type, const flatbuffers::resolver_function_t *resolver) {
+  switch (type) {
+    case MessageType::HandshakeClient: {
+      auto ptr = reinterpret_cast<const HandshakeClient *>(obj);
+      return ptr->UnPack(resolver);
+    }
+    case MessageType::Commands: {
+      auto ptr = reinterpret_cast<const Commands *>(obj);
+      return ptr->UnPack(resolver);
+    }
+    case MessageType::HandshakeServer: {
+      auto ptr = reinterpret_cast<const HandshakeServer *>(obj);
+      return ptr->UnPack(resolver);
+    }
+    case MessageType::FrameState: {
+      auto ptr = reinterpret_cast<const FrameState *>(obj);
+      return ptr->UnPack(resolver);
+    }
+    case MessageType::PlayerLeft: {
+      auto ptr = reinterpret_cast<const PlayerLeft *>(obj);
+      return ptr->UnPack(resolver);
+    }
+    case MessageType::EndGame: {
+      auto ptr = reinterpret_cast<const EndGame *>(obj);
+      return ptr->UnPack(resolver);
+    }
+    case MessageType::Error: {
+      auto ptr = reinterpret_cast<const Error *>(obj);
+      return ptr->UnPack(resolver);
+    }
+    default: return nullptr;
+  }
+}
+
+inline flatbuffers::Offset<void> MessageTypeUnion::Pack(flatbuffers::FlatBufferBuilder &_fbb, const flatbuffers::rehasher_function_t *_rehasher) const {
+  switch (type) {
+    case MessageType::HandshakeClient: {
+      auto ptr = reinterpret_cast<const HandshakeClientT *>(value);
+      return CreateHandshakeClient(_fbb, ptr, _rehasher).Union();
+    }
+    case MessageType::Commands: {
+      auto ptr = reinterpret_cast<const CommandsT *>(value);
+      return CreateCommands(_fbb, ptr, _rehasher).Union();
+    }
+    case MessageType::HandshakeServer: {
+      auto ptr = reinterpret_cast<const HandshakeServerT *>(value);
+      return CreateHandshakeServer(_fbb, ptr, _rehasher).Union();
+    }
+    case MessageType::FrameState: {
+      auto ptr = reinterpret_cast<const FrameStateT *>(value);
+      return CreateFrameState(_fbb, ptr, _rehasher).Union();
+    }
+    case MessageType::PlayerLeft: {
+      auto ptr = reinterpret_cast<const PlayerLeftT *>(value);
+      return CreatePlayerLeft(_fbb, ptr, _rehasher).Union();
+    }
+    case MessageType::EndGame: {
+      auto ptr = reinterpret_cast<const EndGameT *>(value);
+      return CreateEndGame(_fbb, ptr, _rehasher).Union();
+    }
+    case MessageType::Error: {
+      auto ptr = reinterpret_cast<const ErrorT *>(value);
+      return CreateError(_fbb, ptr, _rehasher).Union();
+    }
+    default: return 0;
+  }
+}
+
+inline MessageTypeUnion::MessageTypeUnion(const MessageTypeUnion &u) FLATBUFFERS_NOEXCEPT : type(u.type), value(nullptr) {
+  switch (type) {
+    case MessageType::HandshakeClient: {
+      assert(false);  // HandshakeClientT not copyable.
+      break;
+    }
+    case MessageType::Commands: {
+      assert(false);  // CommandsT not copyable.
+      break;
+    }
+    case MessageType::HandshakeServer: {
+      assert(false);  // HandshakeServerT not copyable.
+      break;
+    }
+    case MessageType::FrameState: {
+      assert(false);  // FrameStateT not copyable.
+      break;
+    }
+    case MessageType::PlayerLeft: {
+      value = new PlayerLeftT(*reinterpret_cast<PlayerLeftT *>(u.value));
+      break;
+    }
+    case MessageType::EndGame: {
+      value = new EndGameT(*reinterpret_cast<EndGameT *>(u.value));
+      break;
+    }
+    case MessageType::Error: {
+      value = new ErrorT(*reinterpret_cast<ErrorT *>(u.value));
+      break;
+    }
+    default:
+      break;
+  }
+}
+
+inline void MessageTypeUnion::Reset() {
+  switch (type) {
+    case MessageType::HandshakeClient: {
+      auto ptr = reinterpret_cast<HandshakeClientT *>(value);
+      delete ptr;
+      break;
+    }
+    case MessageType::Commands: {
+      auto ptr = reinterpret_cast<CommandsT *>(value);
+      delete ptr;
+      break;
+    }
+    case MessageType::HandshakeServer: {
+      auto ptr = reinterpret_cast<HandshakeServerT *>(value);
+      delete ptr;
+      break;
+    }
+    case MessageType::FrameState: {
+      auto ptr = reinterpret_cast<FrameStateT *>(value);
+      delete ptr;
+      break;
+    }
+    case MessageType::PlayerLeft: {
+      auto ptr = reinterpret_cast<PlayerLeftT *>(value);
+      delete ptr;
+      break;
+    }
+    case MessageType::EndGame: {
+      auto ptr = reinterpret_cast<EndGameT *>(value);
+      delete ptr;
+      break;
+    }
+    case MessageType::Error: {
+      auto ptr = reinterpret_cast<ErrorT *>(value);
+      delete ptr;
+      break;
+    }
+    default: break;
+  }
+  value = nullptr;
+  type = MessageType::NONE;
+}
+
+inline const torchcraft::fbs::Message *GetMessage(const void *buf) {
+  return flatbuffers::GetRoot<torchcraft::fbs::Message>(buf);
+}
+
+inline Message *GetMutableMessage(void *buf) {
+  return flatbuffers::GetMutableRoot<Message>(buf);
+}
+
+inline bool VerifyMessageBuffer(
     flatbuffers::Verifier &verifier) {
-  return verifier.VerifyBuffer<torchcraft::fbs::Replayer>(nullptr);
+  return verifier.VerifyBuffer<torchcraft::fbs::Message>(nullptr);
 }
 
-inline void FinishReplayerBuffer(
+inline void FinishMessageBuffer(
     flatbuffers::FlatBufferBuilder &fbb,
-    flatbuffers::Offset<torchcraft::fbs::Replayer> root) {
+    flatbuffers::Offset<torchcraft::fbs::Message> root) {
   fbb.Finish(root);
 }
 
-inline std::unique_ptr<ReplayerT> UnPackReplayer(
+inline std::unique_ptr<MessageT> UnPackMessage(
     const void *buf,
     const flatbuffers::resolver_function_t *res = nullptr) {
-  return std::unique_ptr<ReplayerT>(GetReplayer(buf)->UnPack(res));
+  return std::unique_ptr<MessageT>(GetMessage(buf)->UnPack(res));
 }
 
 }  // namespace fbs
 }  // namespace torchcraft
 
-#endif  // FLATBUFFERS_GENERATED_FRAME_TORCHCRAFT_FBS_H_
+#endif  // FLATBUFFERS_GENERATED_TORCHCRAFT_TORCHCRAFT_FBS_H_
